% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserAttributeEdit.R
\name{addAttributeToQuery}
\alias{addAttributeToQuery}
\title{Function to add Attribute to Query}
\usage{
addAttributeToQuery(query, attribute)
}
\arguments{
\item{query}{identify the query object to edit}

\item{attribute}{the attribute to add to the query}
}
\value{
the edited query component
}
\description{
This function edits a expression type class
}
