% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{ageAtEnd}
\alias{ageAtEnd}
\title{Age at era end attribute (for era and period domains)}
\usage{
ageAtEnd(op)
}
\arguments{
\item{op}{an opAttribute object (numeric or integer) defining the age comparison}
}
\value{
An attribute for use in \code{conditionEra()}, \code{drugEra()}, \code{doseEra()}, \code{observationPeriod()}.
}
\description{
Age at era end attribute (for era and period domains)
}
