% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelLoadFn.R
\name{as.AttributeLoad}
\alias{as.AttributeLoad}
\title{A coercion function to convert to a CAPR attribute}
\usage{
as.AttributeLoad(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
a attribute class object
}
\description{
This function takes a saved CAPR attribute json and returns an attribute CAPR R object
}
