% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelCoercionFn.R
\name{as.Circe,Window-method}
\alias{as.Circe,Window-method}
\alias{as.Circe}
\alias{as.Circe,Timeline-method}
\alias{as.Circe,Occurrence-method}
\alias{as.Circe,ObservationWindow-method}
\alias{as.Circe,Limit-method}
\alias{as.Circe,ExpressionType-method}
\alias{as.Circe,Concept-method}
\alias{as.Circe,ConceptSetItem-method}
\alias{as.Circe,ConceptSetExpression-method}
\alias{as.Circe,OpAttribute-method}
\alias{as.Circe,SourceConceptAttribute-method}
\alias{as.Circe,ConceptAttribute-method}
\alias{as.Circe,LogicAttribute-method}
\alias{as.Circe,CorrelatedCriteriaAttribute-method}
\alias{as.Circe,Query-method}
\alias{as.Circe,QueryAttribute-method}
\alias{as.Circe,Count-method}
\alias{as.Circe,CountAttribute-method}
\alias{as.Circe,Group-method}
\alias{as.Circe,GroupAttribute-method}
\alias{as.Circe,DateOffsetEndStrategy-method}
\alias{as.Circe,CustomEraEndStrategy-method}
\alias{as.Circe,CollapseSettings-method}
\alias{as.Circe,CensorWindow-method}
\alias{as.Circe,Component-method}
\title{Coersive function from S4 to S3}
\usage{
\S4method{as.Circe}{Window}(x)

\S4method{as.Circe}{Timeline}(x)

\S4method{as.Circe}{Occurrence}(x)

\S4method{as.Circe}{ObservationWindow}(x)

\S4method{as.Circe}{Limit}(x)

\S4method{as.Circe}{ExpressionType}(x)

\S4method{as.Circe}{Concept}(x)

\S4method{as.Circe}{ConceptSetItem}(x)

\S4method{as.Circe}{ConceptSetExpression}(x)

\S4method{as.Circe}{OpAttribute}(x)

\S4method{as.Circe}{SourceConceptAttribute}(x)

\S4method{as.Circe}{ConceptAttribute}(x)

\S4method{as.Circe}{LogicAttribute}(x)

\S4method{as.Circe}{CorrelatedCriteriaAttribute}(x)

\S4method{as.Circe}{Query}(x)

\S4method{as.Circe}{Count}(x)

\S4method{as.Circe}{Group}(x)

\S4method{as.Circe}{DateOffsetEndStrategy}(x)

\S4method{as.Circe}{CustomEraEndStrategy}(x)

\S4method{as.Circe}{CollapseSettings}(x)

\S4method{as.Circe}{CensorWindow}(x)

\S4method{as.Circe}{Component}(x)
}
\arguments{
\item{x}{a component class object in s4}
}
\value{
the object converted back to s3 that can be used for json seralization
}
\description{
To serialize between json and R, an S3 list object is required. CAPR creates an organized s4 object that maintains
components of the cohort definition. CIRCE needs to be in an S3 structure in R before serializing to json. These
functions maintain consistency between the s3 and s4 data structures
}
