% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelLoadFn.R
\name{as.CohortEra}
\alias{as.CohortEra}
\title{A coercion function to convert to a CAPR CohortEra}
\usage{
as.CohortEra(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
a cohortEra class object
}
\description{
A coercion function to convert to a CAPR CohortEra
}
