% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelLoadFn.R
\name{as.ComponentLoad}
\alias{as.ComponentLoad}
\title{A coercion function to convert to a CAPR component}
\usage{
as.ComponentLoad(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
a component class object
}
\description{
This function takes a saved CAPR component json and returns component CAPR R object
}
