% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelLoadFn.R
\name{as.Concept}
\alias{as.Concept}
\title{A coercion function to convert to a CAPR concept}
\usage{
as.Concept(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
a concept class object
}
\description{
This function takes a data frame containing information about a concept and converts it into the Concept class
}
