% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelLoadFn.R
\name{as.ConceptLoad}
\alias{as.ConceptLoad}
\title{A coercion function to load to a CAPR concept}
\usage{
as.ConceptLoad(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
a concept class object
}
\description{
This function takes a data frame containing information about a concept and converts it into the Concept class
}
