% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelLoadFn.R
\name{as.ConceptSetExpression}
\alias{as.ConceptSetExpression}
\title{A coercion function to convert to a CAPR conceptSetExpression}
\usage{
as.ConceptSetExpression(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
a concept set expression class object
}
\description{
A coercion function to convert to a CAPR conceptSetExpression
}
