% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelLoadFn.R
\name{as.ConceptSetItem}
\alias{as.ConceptSetItem}
\title{A coercion function to convert to a CAPR conceptSetItem}
\usage{
as.ConceptSetItem(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
a conceptSetItem class object
}
\description{
This function takes a list and converts it into the Concept set Item class
}
