% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelLoadFn.R
\name{as.CountLoad}
\alias{as.CountLoad}
\title{A coercion function to convert to a CAPR count}
\usage{
as.CountLoad(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
a count class object
}
\description{
This function takes a saved CAPR count json and returns count CAPR R object
}
