% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSet.R
\name{as.data.frame,ConceptSet-method}
\alias{as.data.frame,ConceptSet-method}
\title{Coerce a concept set expression to a dataframe}
\usage{
\S4method{as.data.frame}{ConceptSet}(x)
}
\arguments{
\item{x}{A Caper Concept Set}
}
\value{
A tibble (dataframe) with columns: concept_id, includeDescendants, isExcluded, includeMapped.
}
\description{
Coerce a concept set expression to a dataframe
}
