% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelLoadFn.R
\name{as.EndStrategyLoad}
\alias{as.EndStrategyLoad}
\title{A coercion function to convert to a CAPR EndStrategy}
\usage{
as.EndStrategyLoad(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
a EndStrategy class object
}
\description{
This function takes a saved CAPR EndStrategy json and returns EndStrategy CAPR R object
}
