% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelLoadFn.R
\name{as.ExpressionType}
\alias{as.ExpressionType}
\title{A coercion function to convert to a CAPR expression type}
\usage{
as.ExpressionType(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
an expressionType class object
}
\description{
A coercion function to convert to a CAPR expression type
}
