% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelLoadFn.R
\name{as.GroupLoad}
\alias{as.GroupLoad}
\title{A coercion function to convert to a CAPR group}
\usage{
as.GroupLoad(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
a group class object
}
\description{
This function takes a saved CAPR group json and returns group CAPR R object
}
