% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{as.json}
\alias{as.json}
\title{Coerce Capr object to json}
\usage{
as.json(x, pretty = TRUE, ...)
}
\arguments{
\item{x}{the capr object}

\item{pretty}{a toggle to make the json look nice, part of jsonlite}

\item{...}{additional arguments passes to jsonlite::toJSON}
}
\description{
Coerce Capr object to json
}
