% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelLoadFn.R
\name{as.Limit}
\alias{as.Limit}
\title{A coercion function to convert to a CAPR limit}
\usage{
as.Limit(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
a limit class object
}
\description{
A coercion function to convert to a CAPR limit
}
