% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelLoadFn.R
\name{as.MetaData}
\alias{as.MetaData}
\title{A coercion function to convert to a CAPR metaData}
\usage{
as.MetaData(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
a meta data class object
}
\description{
A coercion function to convert to a CAPR metaData
}
