% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelLoadFn.R
\name{as.ObservationWindow}
\alias{as.ObservationWindow}
\title{A coercion function to convert to a CAPR ObservationWindow}
\usage{
as.ObservationWindow(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
an observation window class object
}
\description{
A coercion function to convert to a CAPR ObservationWindow
}
