% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelLoadFn.R
\name{as.Occurrence}
\alias{as.Occurrence}
\title{A coercion function to convert to a CAPR Occurrence}
\usage{
as.Occurrence(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
a occurrence class object
}
\description{
A coercion function to convert to a CAPR Occurrence
}
