% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelLoadFn.R
\name{as.QueryLoad}
\alias{as.QueryLoad}
\title{A coercion function to convert to a CAPR query}
\usage{
as.QueryLoad(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
a query class object
}
\description{
This function takes a saved CAPR query json and returns query CAPR R object
}
