% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelLoadFn.R
\name{as.Timeline}
\alias{as.Timeline}
\title{A coercion function to convert to a CAPR timeline}
\usage{
as.Timeline(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
a timeline class object
}
\description{
A coercion function to convert to a CAPR timeline
}
