% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelLoadFn.R
\name{as.Window}
\alias{as.Window}
\title{A coercion function to convert to a CAPR window}
\usage{
as.Window(x)
}
\arguments{
\item{x}{the object to coerce}
}
\value{
a window class object
}
\description{
A coercion function to convert to a CAPR window
}
