% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{atMost}
\alias{atMost}
\title{Require at most N occurrences of a query within a window}
\usage{
atMost(
  x,
  query,
  aperture = duringInterval(eventStarts(-Inf, Inf)),
  distinct = NA,
  countColumn = NA_character_
)
}
\arguments{
\item{x}{Maximum integer count of allowed occurrences.}

\item{query}{A \code{Query} object (e.g. from \code{conditionOccurrence()}).}

\item{aperture}{An \code{EventAperture} built with \code{\link{duringInterval}}. Defaults to all time.}

\item{distinct}{If \code{TRUE}, count distinct values of \code{countColumn}.}

\item{countColumn}{Column to use for distinct counting (e.g. \code{"DOMAIN_CONCEPT"}).}
}
\value{
A \code{Criteria} object for use in \code{\link{withAll}}, \code{\link{withAny}}, etc.
}
\description{
Require at most N occurrences of a query within a window
}
\examples{
t2dm <- cs(descendants(201826L), name = "T2DM")
# At most 1 T2DM diagnosis before index
atMost(1, conditionOccurrence(t2dm), duringInterval(eventStarts(-365, 0)))
}
\seealso{
\code{\link{exactly}}, \code{\link{atLeast}}, \code{\link{duringInterval}}
}
