% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter.R
\name{cap_adapter}
\alias{cap_adapter}
\title{Resolve an adapter through S3}
\usage{
cap_adapter(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Reserved for bridge methods.}
}
\value{
An adapter or \code{NULL}. Methods must not run digest orchestration.
}
\description{
Resolve an adapter through S3
}
