% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{cap_pack_conformance_report}
\alias{cap_pack_conformance_report}
\title{Emit a pack conformance report}
\usage{
cap_pack_conformance_report(pack)
}
\arguments{
\item{pack}{Hosted pack.}
}
\value{
A \code{cap.pack_conformance_report.v1}.
}
\description{
Emit a pack conformance report
}
