% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{cap_registry}
\alias{cap_registry}
\title{Create or access an adapter registry}
\usage{
cap_registry(global = TRUE)
}
\arguments{
\item{global}{If \code{TRUE}, return the process-global registry. Otherwise
create a new isolated registry.}
}
\value{
A mutable \code{capr_registry} environment.
}
\description{
Create or access an adapter registry
}
