% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{cap_unregister_adapter}
\alias{cap_unregister_adapter}
\title{Unregister adapters}
\usage{
cap_unregister_adapter(
  class = NULL,
  id = NULL,
  version = NULL,
  registry = cap_registry()
)
}
\arguments{
\item{class}{Optional host class.}

\item{id}{Optional adapter ID.}

\item{version}{Optional adapter version.}

\item{registry}{Registry returned by \code{cap_registry()}.}
}
\value{
The number of removed entries.
}
\description{
Unregister adapters
}
