% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter.R
\name{cap_validate_adapter}
\alias{cap_validate_adapter}
\title{Validate an adapter}
\usage{
cap_validate_adapter(adapter)
}
\arguments{
\item{adapter}{An adapter object.}
}
\value{
The adapter, invisibly.
}
\description{
Validate an adapter
}
