% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{cap_validate_pack}
\alias{cap_validate_pack}
\title{Validate hosted pack metadata}
\usage{
cap_validate_pack(pack)
}
\arguments{
\item{pack}{Pack from \code{cap_load_pack()}.}
}
\value{
Fixture-compatible pack validation.
}
\description{
Validate hosted pack metadata
}
