% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capr.R
\docType{package}
\name{Capr-package}
\alias{Capr}
\alias{Capr-package}
\title{Capr: Cohort Definition Application Programming}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a programming language for defining cohort definitions in R to use in studies for Observational Health Data Sciences and Informatics (OHDSI). The functions in 'Capr' allow for the programmatic creation of OHDSI concept sets and cohorts that can be serialized to 'OHDSI' compatible 'json' files or to 'OHDSI-SQL'. 'Capr' functions can be used to create, save, and load component parts to a cohort definition allowing R programmers to easily reuse cohort logic. 'Capr' provides tools to create a large number of OHDSI cohorts while also helping bridge the gap between human readable descriptions of clinical phenotypes and their computational implementation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/Capr/}
  \item \url{https://github.com/OHDSI/Capr/}
  \item Report bugs at \url{https://github.com/OHDSI/Capr/issues/}
}

}
\author{
\strong{Maintainer}: Martin Lavallee \email{martin.lavallee@boehringer-ingelheim.com}

Authors:
\itemize{
  \item Martin Lavallee \email{martin.lavallee@boehringer-ingelheim.com}
  \item Adam Black \email{black@ohdsi.org}
}

Other contributors:
\itemize{
  \item Katy Sadowski \email{sadowski@ohdsi.org} [contributor]
}

}
\keyword{internal}
