% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capr.R
\name{capr}
\alias{capr}
\title{Covariate Assisted Projection (CAP) Regression}
\usage{
capr(
  S,
  X,
  K,
  B.init = NULL,
  Gamma.init = NULL,
  weight = NULL,
  max_iter = 200L,
  tol = 1e-06,
  orth = TRUE,
  n.init = 10L
)
}
\arguments{
\item{S}{Numeric 3D array of size \eqn{p \times p \times n} (for example, stack of covariance matrices).}

\item{X}{Numeric matrix \eqn{n \times q} (design matrix), created for example by \code{model.matrix}.}

\item{K}{Integer scalar, number of components (\eqn{K \ge 1}).}

\item{B.init}{Initial value of coefficient \eqn{B \in R^{q \times n.init \times K}} (default zero 3D array).}

\item{Gamma.init}{Initial value of principal direction matrix  \eqn{\Gamma \in  R^{p\times n.init \times K}} (default random Gaussian 3D array).}

\item{weight}{Numeric vector of length \eqn{n} (default rep(1, n)), where each element should be proportional the sample size for each \eqn{S[,,i]}.}

\item{max_iter}{Integer scalar, max flip–flop iterations per component (default 200).}

\item{tol}{Positive numeric scalar, convergence tolerance (default 1e-6).}

\item{orth}{Logical scalar; if TRUE (default), enforce orthogonality (recommended) of successive \eqn{\gamma}. If FALSE, no orthogonalization is performed with warnings of potential identical components.}

\item{n.init}{Integer scalar; number of random initializations (default 10). If B.init and Gamma.init are both set properly, n.init is ignored.}
}
\value{
A list in class capr with:
\item{B}{numeric matrix \eqn{q \times K}, column \eqn{k} stores \eqn{\beta^{(k)}}}
\item{Gamma}{numeric matrix \eqn{p \times K}, column \eqn{k} stores \eqn{\gamma^{(k)}}}
\item{loglike}{negative loglikehood, upto constant scaling and shift}
\item{S}{3D array used for fitting}
\item{X}{design matrix used for fitting}
\item{weight}{weight values used for fitting}
}
\description{
Fits (multiple) CAP components sequentially for each of the
principal direction vectors \eqn{\gamma^{(k)}} and corresponding regression coefficients
\eqn{\beta^{(k)}}, \eqn{k = 1, \dots, K}. Each component is estimated via a flip–flop algorithm,
with recommended orthogonalization of successive directions.
}
\details{
The CAP negative log-likelihood is defined as \deqn{  	\underset{\boldsymbol{\beta},\boldsymbol{\gamma}}{\text{minimize}} &&    \frac{1}{2}\sum_{i=1}^{n}(\bx_{i}^\top\boldsymbol{\beta}^{(k)}) \cdot T_{i}  +\frac{1}{2}\sum_{i=1}^{n} \boldsymbol{\gamma}^{(k)\top} \bS_{i}^{(k)} \boldsymbol{\gamma}^{(k)} \cdot \exp(-\bx_{i}^\top\boldsymbol{\beta}^{(k)}) , \nonumber \\
The algoirhtm fit the \eqn{\gamma} and \eqn{\beta} sequentially with (ranom) multiple  innitializations  and returns the  solution pair that minimizes the negative log-likehood.
\end{eqnarray}} for each component.
}
\examples{
simu.data <- simu.capr(seed = 123L, n = 120L)
K <- 2L
fit <- capr(
    S = simu.data$S,
    X = simu.data$X,
    K = K
)
print(fit)
}
\references{
Zhao, Y., Wang, B., Mostofsky, S. H., Caffo, B. S., & Luo, X. (2021). Covariate assisted principal regression for covariance matrix outcomes. Biostatistics, 22(3), 629-645.
}
