% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capr_canonical_json}
\alias{capr_canonical_json}
\title{Serialize deterministic JSON}
\usage{
capr_canonical_json(x, pretty = FALSE)
}
\arguments{
\item{x}{An R object representable as JSON.}

\item{pretty}{Whether to indent the output.}
}
\value{
A UTF-8 JSON string with deterministically ordered object keys.
}
\description{
Serialize deterministic JSON
}
