% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit.R
\docType{class}
\name{CensoringCriteria-class}
\alias{CensoringCriteria-class}
\title{An S4 class identifying a censoring criteria for the cohort}
\description{
The censoring criteria specifies events where the person exits the cohort.
These events are based on a query class object and users can specify multiple queries
in the censoring criteria.
}
\section{Slots}{

\describe{
\item{\code{criteria}}{a list of Capr query class objects that specify the events that
would lead a person to exit the cohort.}
}}

