% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit.R
\name{censoringEvents}
\alias{censoringEvents}
\title{Constructor for a set of censoring events}
\usage{
censoringEvents(...)
}
\arguments{
\item{...}{a list of Capr query objects that are used as censoring events}
}
\value{
a censoring criteria S4 object used to define the censoring events of the cohort definition
}
\description{
Constructor for a set of censoring events
}
