% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelClasses.R
\docType{class}
\name{CensorWindow-class}
\alias{CensorWindow-class}
\title{An S4 class for CensorWindow}
\description{
A class showing dates that indicate the range of entries the are captured in the cohort
}
\section{Slots}{

\describe{
\item{\code{StartDate}}{the left side of truncation for the study observation}

\item{\code{EndDate}}{the right side of truncation for the study observation}
}}

