% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelUtilityFn.R
\name{checkConceptField}
\alias{checkConceptField}
\title{Function to get concept fields from concept set expression in object}
\usage{
checkConceptField(x, field)
}
\arguments{
\item{x}{the object to check}

\item{field}{the concept field to check}
}
\value{
a list or vector of concept fields
}
\description{
Function to get concept fields from concept set expression in object
}
