% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelUtilityFn.R
\name{checkConceptIds}
\alias{checkConceptIds}
\title{Function to get concept ids from concept set expression in object}
\usage{
checkConceptIds(x)
}
\arguments{
\item{x}{the object to check}
}
\value{
a list or vector of concept id integers
}
\description{
Function to get concept ids from concept set expression in object
}
