% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\docType{class}
\name{CohortDetails-class}
\alias{CohortDetails-class}
\title{Show Contents of a Component}
\description{
This function prints the contents of a component. Note 1/27/21 attributes and some other s4 classes
need to be implemented
}
\details{
param showFullConceptSetExpressions T/F options to include full details of concept expressions
An S4 class providing metadata for a CohortDefinition

The cohort details do not affect the cohort definition and are for improving
human readability only.
}
\section{Slots}{

\describe{
\item{\code{Name}}{a name for the cohort}

\item{\code{Description}}{a text field providing an information on the cohort and what it is intended}

\item{\code{Author}}{who created the cohort}

\item{\code{cdmVersionRange}}{the range of cdm versions}
}}

