% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\docType{class}
\name{CollapseSettings-class}
\alias{CollapseSettings-class}
\title{An S4 class for Collapse Settings}
\description{
A class providing information that identifies the padding for cohort eras
}
\section{Slots}{

\describe{
\item{\code{Type}}{boolean operator for the number of items in group to include. all, any, at most and at least}

\item{\code{Count}}{the number of criteria's needed for restriction. If Type is ALL or ANY this value is NA}
}}

