% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{compile.Cohort}
\alias{compile.Cohort}
\title{Compile a Capr cohort to json}
\usage{
compile.Cohort(object, ...)
}
\arguments{
\item{object}{A Capr cohort or list of Capr cohorts}

\item{...}{Arguments passed on to jsonlite::toJSON.
e.g. `pretty = TRUE` for nicely formatted json.}
}
\value{
The json representation of Capr cohorts
}
\description{
Compile a Capr cohort to json
}
\examples{
\dontrun{
ch <- cohort(condition(cs(1,2)))
compile(ch)
}
}
