% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{compile}
\alias{compile}
\title{Compile a Capr object to json}
\usage{
compile(object, ...)
}
\arguments{
\item{object}{A Capr object such as a cohort, list of cohorts, or concept set.}

\item{...}{Arguments passed on to jsonlite::toJSON (e.g. \code{pretty = TRUE}).
For the Cohort method, \code{includeConceptSets} is also allowed; see
\code{\link{compile,Cohort-method}}.}
}
\value{
The json representation of the Capr object
}
\description{
Compile a Capr object to json
}
