% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCommands.R
\name{compileCohortDefinition}
\alias{compileCohortDefinition}
\title{Convert cohort definition object to CIRCE and run through circe compiler}
\usage{
compileCohortDefinition(CohortDefinition, generateOptions)
}
\arguments{
\item{CohortDefinition}{input cohort Definition class object}

\item{generateOptions}{the options for building the ohdisql using CirceR::createGenerateOptions}
}
\value{
A three tiered list containing the circe converted cohort definition, the circe json and ohisql.
If an error occurs the ohdisql slot will be NA and the user should review the circe cohort definition for potential errors.
}
\description{
This function converts a Cohort Definition class object to a CIRCE expression, creates the json and compiles the
circe json to create ohdisql to run queries against a dbms containing OMOP cdm data
}
