% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelUtilityFn.R
\name{componentClass,Component-method}
\alias{componentClass,Component-method}
\alias{componentClass}
\title{Function to find the Component Class}
\usage{
\S4method{componentClass}{Component}(x)
}
\arguments{
\item{x}{the component to check}
}
\value{
a character string with the component class
}
\description{
Function to find the Component Class
}
