% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelUtilityFn.R
\name{componentType,Component-method}
\alias{componentType,Component-method}
\alias{componentType}
\title{Function to find the Component Class}
\usage{
\S4method{componentType}{Component}(x)
}
\arguments{
\item{x}{the component to check}
}
\value{
a character string with the component class
}
\description{
Function to find the Component Class
}
