% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\docType{class}
\name{Concept-class}
\alias{Concept-class}
\title{An S4 class for a Concepet}
\description{
A concept class contains all the information about the concept from the OMOP voabulary
}
\section{Slots}{

\describe{
\item{\code{CONCEPT_ID}}{the id of the concept}

\item{\code{CONCEPT_NAME}}{the name of the concept}

\item{\code{STANDARD_CONCEPT}}{whether the cncept is standard, single letter}

\item{\code{STANDARD_CONCEPT_CAPTION}}{whether the concept is standard full phrase}

\item{\code{INVALID_REASON}}{Whether the concept is invalid single letter}

\item{\code{INVALID_REASON_CAPTION}}{whether the concept is invalid standard phrase}

\item{\code{CONCEPT_CODE}}{the original code of the concept from its vocabulary}

\item{\code{DOMAIN_ID}}{the domain of the concept}

\item{\code{VOCABULARY_ID}}{the name of the vocabulary}

\item{\code{CONCEPT_CLASS_ID}}{type of concept class}
}}

