% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelClasses.R
\docType{class}
\name{ConceptAttribute-class}
\alias{ConceptAttribute-class}
\title{An S4 class for Concept Attribute}
\description{
A concept attribute, using concepts to identify the attribute like a gender or race etc
}
\section{Slots}{

\describe{
\item{\code{Name}}{the name of the attribute}

\item{\code{Concepts}}{a list containing the concepts used to identify the attribute}
}}

