% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\docType{class}
\name{conceptSetAttribute-class}
\alias{conceptSetAttribute-class}
\title{An S4 class for a concept set attribute that holds a reference to a concept set ID}
\description{
An S4 class for a concept set attribute that holds a reference to a concept set ID
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the attribute}

\item{\code{conceptSet}}{a ConceptSet object that provides the ID reference}
}}

