% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\docType{class}
\name{ConceptSetExpression-class}
\alias{ConceptSetExpression-class}
\title{An S4 class for ConceptSetExpresion}
\description{
A class for the concept set expressions bundles multiple concepts with mapping
}
\section{Slots}{

\describe{
\item{\code{id}}{an id for the concept set expression to identify within a component}

\item{\code{Name}}{the name of the concept set expression}

\item{\code{Expression}}{a list containing expressions. expressions include multiple conceptSetItem objects}
}}

