% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\docType{class}
\name{ConceptSetItem-class}
\alias{ConceptSetItem-class}
\title{An S4 class for ConceptSetItem}
\description{
a class that provides information on the mapping of the concept
}
\section{Slots}{

\describe{
\item{\code{Concept}}{a concept class object}

\item{\code{isExcluded}}{toggle if want to exclude the concept}

\item{\code{includeDescendants}}{toggle if want to include descendants}

\item{\code{includeMapped}}{toggle if want to include map}
}}

