% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{condition}
\alias{condition}
\title{Query the condition domain}
\usage{
condition(conceptSet, ...)
}
\arguments{
\item{conceptSet}{A condition concept set}

\item{...}{optional attributes}
}
\value{
A Capr Query
}
\description{
Query the condition domain
}
