% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{conditionSourceConcept}
\alias{conditionSourceConcept}
\title{Add a condition source concept attribute}
\usage{
conditionSourceConcept(conceptSet)
}
\arguments{
\item{conceptSet}{a ConceptSet object containing the source concepts}
}
\value{
An attribute that can be used in a query function
}
\description{
Add a condition source concept attribute
}
