% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-conceptSetSelection.R
\name{conditionStatusCS}
\alias{conditionStatusCS}
\title{Condition status filter via concept set — ConditionOccurrence}
\usage{
conditionStatusCS(conceptSet, isExclusion = FALSE)
}
\arguments{
\item{conceptSet}{A \code{ConceptSet} built with \code{cs()}.}

\item{isExclusion}{logical; if \code{TRUE} the types in the concept set are excluded.}
}
\value{
An attribute for use in \code{conditionOccurrence()}.
}
\description{
Condition status filter via concept set — ConditionOccurrence
}
