% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{contains}
\alias{contains}
\alias{contains,character-method}
\title{String contains operator}
\usage{
contains(x)

\S4method{contains}{character}(x)
}
\arguments{
\item{x}{the substring to search for}
}
\description{
function that builds an opAttribute based on substring containment logic
}
