% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{continuousObservation}
\alias{continuousObservation}
\title{A function to construct the observationWindow}
\usage{
continuousObservation(priorDays = 0L, postDays = 0L)
}
\arguments{
\item{priorDays}{minimum number of observation days prior to the cohort index. Default 0 days}

\item{postDays}{minimum number of observation days post cohort index. Default 0 days}
}
\description{
A function to construct the observationWindow
}
