% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelCoercionFn.R
\name{convertCensoringCriteriaToCIRCE}
\alias{convertCensoringCriteriaToCIRCE}
\title{Convert Censoring Criteria Component to CIRCE}
\usage{
convertCensoringCriteriaToCIRCE(x)
}
\arguments{
\item{x}{the component to convert}
}
\value{
a circe converted component
}
\description{
Convert Censoring Criteria Component to CIRCE
}
