% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelCoercionFn.R
\name{convertCohortDefinitionToCIRCE}
\alias{convertCohortDefinitionToCIRCE}
\title{Function to update cohort definition to CIRCE}
\usage{
convertCohortDefinitionToCIRCE(x)
}
\arguments{
\item{x}{the cohort definition to convert to circe}
}
\value{
a circe object in R
}
\description{
Function to update cohort definition to CIRCE
}
