% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelCoercionFn.R
\name{convertCohortEraToCIRCE}
\alias{convertCohortEraToCIRCE}
\title{Convert CohortEra Component to CIRCE}
\usage{
convertCohortEraToCIRCE(x)
}
\arguments{
\item{x}{the component to convert}
}
\value{
a circe converted component
}
\description{
Convert CohortEra Component to CIRCE
}
