% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelCoercionFn.R
\name{convertEndStrategyToCIRCE}
\alias{convertEndStrategyToCIRCE}
\title{Convert End Strategy Component to CIRCE}
\usage{
convertEndStrategyToCIRCE(x)
}
\arguments{
\item{x}{the component to convert}
}
\value{
a circe converted component
}
\description{
Convert End Strategy Component to CIRCE
}
