% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelCoercionFn.R
\name{convertInclusionRulesToCIRCE}
\alias{convertInclusionRulesToCIRCE}
\title{Convert Inclusion Rules Component to CIRCE}
\usage{
convertInclusionRulesToCIRCE(x)
}
\arguments{
\item{x}{the component to convert}
}
\value{
a circe converted component
}
\description{
Convert Inclusion Rules Component to CIRCE
}
