% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelCoercionFn.R
\name{convertPrimaryCriteriaToCIRCE}
\alias{convertPrimaryCriteriaToCIRCE}
\title{Convert Primary Criteria Component to CIRCE}
\usage{
convertPrimaryCriteriaToCIRCE(x)
}
\arguments{
\item{x}{the component to convert}
}
\value{
a circe converted component
}
\description{
Convert Primary Criteria Component to CIRCE
}
