% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelCoercionFn.R
\name{convertRuleToCIRCE}
\alias{convertRuleToCIRCE}
\title{Convert single rule (group) Component to CIRCE}
\usage{
convertRuleToCIRCE(x)
}
\arguments{
\item{x}{the component to convert}
}
\value{
a circe converted component
}
\description{
Convert single rule (group) Component to CIRCE
}
