% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelClasses.R
\docType{class}
\name{CorrelatedCriteriaAttribute-class}
\alias{CorrelatedCriteriaAttribute-class}
\title{An S4 class for CorrelatedCriteriaAttribue}
\description{
A group attribute that is nested within a query.
}
\section{Slots}{

\describe{
\item{\code{Name}}{name of the attribute}

\item{\code{Group}}{a group class object for the attribute}
}}

