% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelClasses.R
\docType{class}
\name{Count-class}
\alias{Count-class}
\title{An S4 class for a Count}
\description{
A count class provides a number of occurrences of the query and the timeline that it happens
}
\section{Slots}{

\describe{
\item{\code{Criteria}}{a query class object}

\item{\code{Timeline}}{a timeline class object}

\item{\code{Occurrence}}{an occurrence class object}
}}

