% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateFn.R
\name{createAdditionalCriteria}
\alias{createAdditionalCriteria}
\title{Function creates an Additional Criteria}
\usage{
createAdditionalCriteria(Name, Contents = NULL, Limit, Description = NULL)
}
\arguments{
\item{Name}{a character string naming the group object, this is required for the object. One should make the name
descriptive of what the group is trying to identify.}

\item{Contents}{a single component of group class that describes the additional criteria. If the Contents
are empty then the additional criteria is only decribed by the qualified limit}

\item{Limit}{how to limit initial events per person}

\item{Description}{a character string describing the count object, this is optional so default is null}
}
\value{
new additional criteria component.
}
\description{
Function creates an Additional Criteria from a component class group
}
