% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateFn.R
\name{createCensoringCriteria}
\alias{createCensoringCriteria}
\title{Function creates a Censoring Criteria}
\usage{
createCensoringCriteria(Name, ComponentList, Description = NULL)
}
\arguments{
\item{Name}{a character string naming the inclusion rules, this is required for the object. One should make the name
descriptive of what the group is trying to identify.}

\item{ComponentList}{a list of component class queries to be inserted into the censoring criteria.}

\item{Description}{a character string describing the count object, this is optional so default is null}
}
\value{
new censoring criteria component.
}
\description{
Function creates a Censoring Criteria from a list of queries
}
