% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCommands.R
\name{createCohortDataframe}
\alias{createCohortDataframe}
\title{Create a dataframe that can be used by cohort generator to build cohorts into backend table}
\usage{
createCohortDataframe(cohortList, generateStats = TRUE)
}
\arguments{
\item{cohortList}{a list of cohorts to turn into a dataframe}

\item{generateStats}{select true if you want inclusion rule stats or false for no stats}
}
\value{
A data frame of cohort definitions containing meta data
}
\description{
This function converts a list of Cohort definitions into a dataframe that can be used by cohort
generator and cohort diagnostics to build and evaluate cohorts. The dataframe returns an atlasId column and
a cohortId column which are equivalent. Further it creates a column for the ohdisql, json, read and name of
cohort.
}
