% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userCreateFn.R
\name{createCohortDefinition}
\alias{createCohortDefinition}
\title{Create Cohort Definition class object}
\usage{
createCohortDefinition(
  Name,
  Description = NA_character_,
  Author = NA_character_,
  cdmVersionRange = ">=5.0.0",
  PrimaryCriteria,
  AdditionalCriteria = NULL,
  InclusionRules = NULL,
  EndStrategy = NULL,
  CensoringCriteria = NULL,
  CohortEra = NULL
)
}
\arguments{
\item{Name}{make a name for the cohort to add to the cohort details}

\item{Description}{add a description detail to cohort details, optional}

\item{Author}{add an author name to cohort details, optional}

\item{cdmVersionRange}{add a cdm version range typically >= 5.0.0, please specify if not v5}

\item{PrimaryCriteria}{add primary criteria object}

\item{AdditionalCriteria}{add additional criteria object. if null then will create an additional criteria with qualified limit}

\item{InclusionRules}{add inclusion rules object. if null will create empty inclusion rules with expression limit}

\item{EndStrategy}{add end strategy object. if null will add end of continuous era strategy}

\item{CensoringCriteria}{add censoring criteria object. if null will add empty censoring criteria}

\item{CohortEra}{add cohort era object. if null will add collapse settings with 0 day pad and no censor window}
}
\value{
cohort definition class object with defined inputs. This can now be compiled into ohdisql and converted to json
}
\description{
This function creates a Cohort Definition class object from multiple component parts. A cohort definition
contains at a minimum a primary criteria class. The cohort definition can further contain a inclusion rules,
additional criteria, censoring criteria and end strategy classes to provide more details on cohort restriction
and cohort exit. Other components may also be manipulated but since they do not rely on a concept set expressions,
they can be manipulated in separate methods. The cohort definition class differs from the circe expression in that
it does not have a separate space for concept set expressions, which are bundled within the component.
}
