% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userCreateFn.R
\name{createCohortEra}
\alias{createCohortEra}
\title{Create a Cohort Era class object}
\usage{
createCohortEra(EraPadDays = 0L, LeftCensorDate = NULL, RightCensorDate = NULL)
}
\arguments{
\item{EraPadDays}{a numeric that specifies the number of days for the era padding}

\item{LeftCensorDate}{a date string that specifies the starting date of registration}

\item{RightCensorDate}{a date string that specifies the end date of registration}
}
\value{
a cohort era component
}
\description{
The Cohort Era depicts the time span of the cohort. The Censor Window includes
the date window for which we register events. The Collapse Settings identify the era padding
between events before exiting a cohort.
}
