% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelCreateFn.R
\name{createComponent}
\alias{createComponent}
\title{createComponent}
\usage{
createComponent(
  Name,
  Description = NULL,
  ComponentClass = c("ConceptSetExpression", "Group", "Query", "Count", "Attribute",
    "PrimaryCriteria", "AdditionalCriteria", "InclusionRules", "EndStrategy",
    "CensoringCriteria", "CohortEra", "Empty"),
  CriteriaExpression = NULL,
  Limit = NULL,
  ConceptSetExpression = NULL
)
}
\arguments{
\item{Name}{a name}

\item{Description}{a description default null}

\item{ComponentClass}{match an arg from vector}

\item{CriteriaExpression}{include anything for the criteria can be null}

\item{Limit}{determine limit}

\item{ConceptSetExpression}{add anny concept set expressions}
}
\description{
createComponent
}
