% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelCreateFn.R
\name{createConceptAttribute}
\alias{createConceptAttribute}
\title{createConceptAttribue}
\usage{
createConceptAttribute(conceptIds, mapToStandard = TRUE, name)
}
\arguments{
\item{conceptIds}{the list of ids to lookup, need OMOP vocabulary connection}

\item{mapToStandard}{whether to map concept ids to standard or leave as is default is TRUE}

\item{name}{is the name of the attribute}
}
\description{
createConceptAttribue
}
