% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateFn.R
\name{createConceptMapping}
\alias{createConceptMapping}
\title{Function to help user develop the concept mapping}
\usage{
createConceptMapping(
  n,
  includeDescendants = NULL,
  isExcluded = NULL,
  includeMapped = NULL
)
}
\arguments{
\item{n}{the length of the concept set expression}

\item{includeDescendants}{a logic vector of length n that contains the toggle for whether the concept should
include descendants. If the parameter is left null then will return all FALSE}

\item{isExcluded}{a logic vector of length n that contains the toggle for whether the concept should
be excluded. If the parameter is left null then will return all FALSE}

\item{includeMapped}{a logic vector of length n that contains the toggle for whether the concept should
include mapped concepts. If the parameter is left null then will return all FALSE}
}
\value{
This function returns a list for concept mapping for the concept set expression
}
\description{
This function creates a concept mapping list that is used to establish the concept set item for each
member of the concept set expression. This function is evolving.
}
