% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateFn.R
\name{createConceptSetExpression}
\alias{createConceptSetExpression}
\title{Create Concept Set Expression}
\usage{
createConceptSetExpression(
  conceptSet,
  Name,
  includeDescendants = TRUE,
  isExcluded = FALSE,
  includeMapped = FALSE
)
}
\arguments{
\item{conceptSet}{a dataframe containing the concepts one would like to add to the concept set. The data frame
of concepts can be queried using the lookup concept functions (requires a connection to an OMOP CDM).}

\item{Name}{a name for the concept set expression.}

\item{includeDescendants}{logic toggle where default true includes descendant concepts to the defined concept}

\item{isExcluded}{logic toggle when true excludes the defined concept when attached to a concept set expression}

\item{includeMapped}{logic toggle when true includes mapped concepts to the defined concept}
}
\value{
This function returns a component class object which contains the concept set expression
}
\description{
This function takes a data frame of OMOP concepts, establishes the mapping logic and bundles them together as
a concept set expression. A new concept expression created in R sets a guid (unique id) for the concept id. This unique identifier
is used to link the concept set expressions to its implementation within the cohort definition (typically as a query).
With this function, toggling the mapping options sets the logic for all concepts in the concept set expression. If the
user wants to set a custom mapping for each concept in the expression the user should use createConceptSetExpressionCustom.
This is an evolving function.
}
