% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateFn.R
\name{createConceptSetExpressionCustom}
\alias{createConceptSetExpressionCustom}
\title{Create a Custom Concept Set Expression}
\usage{
createConceptSetExpressionCustom(conceptSet, Name, conceptMapping = NULL)
}
\arguments{
\item{conceptSet}{a dataframe containing the concepts one would like to add to the concept set. The data frame
of concepts can be queried using the lookup concept functions (requires a connection to an OMOP CDM).}

\item{Name}{a name for the concept set expression.}

\item{conceptMapping}{a list of mapping for each concept set item. The list will contain whether the concept should
includeDescendants, isExcluded or includeMapped. If the concept Mapping is left null then by default only the
includeDescendants mapping will be true for all. others will remain false.}
}
\value{
This function returns a component class object which contains the concept set expression
}
\description{
This function takes a data frame of OMOP concepts, establishes the mapping logic and bundles them together as
a concept set expression. A new concept expression created in R sets a guid for the concept id. This unique identifier
is used to link the concept set expressions to its implementation within the cohort definition (typically as a query).
With this function, the user can pre-define a full list of mapping for each concept set item in the concept set expression.
This is an evolving function
}
