% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userCreateAttributeFn.R
\name{createConditionTypeExcludeAttribute}
\alias{createConditionTypeExcludeAttribute}
\title{create exclude attribute for condition type}
\usage{
createConditionTypeExcludeAttribute(logic = FALSE)
}
\arguments{
\item{logic}{toggle FALSE to not exclude}
}
\value{
a component of attribute class
}
\description{
This function creates a attribute for exclusion
}
