% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userCreateAttributeFn.R
\name{createCorrelatedCriteriaAttribute}
\alias{createCorrelatedCriteriaAttribute}
\title{Function to create an attribute for a correlated criteria}
\usage{
createCorrelatedCriteriaAttribute(Group)
}
\arguments{
\item{Group}{a group object to add}
}
\value{
a correlated criteria attribute component
}
\description{
Function to create an attribute for a correlated criteria
}
