% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userCreateFn.R
\name{createCount}
\alias{createCount}
\title{Function creates a count object}
\usage{
createCount(
  Query,
  Logic = c("at_least", "at_most", "exactly"),
  Count,
  isDistinct = FALSE,
  Timeline,
  Name = NULL,
  Description = NULL
)
}
\arguments{
\item{Query}{a component that is of query class}

\item{Logic}{how to express the count i.e. exactly, at_least, at_most}

\item{Count}{how many times the query occurs to be eligible}

\item{isDistinct}{a logic toggle where if TRUE only counts distinct occurrences}

\item{Timeline}{a timeline class object orienting the time points of recording in reference to the initial event}

\item{Name}{a character string naming the count object, this is optional so default is null}

\item{Description}{a character string describing the count object, this is optional so default is null}
}
\value{
This function returns a component class object which contains the count object and attached concept
set expressions
}
\description{
This function creates a count object of the cohort definition. The count object is used to express a query over
a number of occurrences within a timeline relative to the initial event. A count comes from the number of times
the applied query must be counted in the candidate patient timeline for them to be a suitable occurrence of a clinical
construct.
}
