% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelBuildLangFn.R
\name{createCountCall}
\alias{createCountCall}
\title{Get counts from cohort expression and prepare R language}
\usage{
createCountCall(x, nm)
}
\arguments{
\item{x}{the circe cohort definition}

\item{nm}{the naming convention to assign the object}
}
\value{
r language to generate the counts of the cohort
}
\description{
Get counts from cohort expression and prepare R language
}
