% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateFn.R
\name{createCustomEraEndStrategy}
\alias{createCustomEraEndStrategy}
\title{Function creates an end strategy from a custom era}
\usage{
createCustomEraEndStrategy(ConceptSetExpression, gapDays, offset)
}
\arguments{
\item{ConceptSetExpression}{a component of concept set expression class that contains information
on the drug concets to use to define the end strategy}

\item{gapDays}{the maximum allowable days between successive exposures.}

\item{offset}{an integer value specifying padding to the cohort exit.}
}
\value{
This function returns a component class object which contains the end strategy object
}
\description{
This function creates a custom era end strategy. From the ATLAS page: Specify a concept set
that contains one or more drugs. A drug era will be derived from all drug exposure events for
any of the drugs within the concept set, using the specified persistence window as a maximum allowable
gap in days between successive exposure events and adding a specified surveillance window to the final exposure event.
 If no exposure event end date is provided, then an exposure event end date is inferred to be event start
 date + days supply in cases when days supply is available or event start date + 1 day otherwise.
This event persistence assures that the cohort end date will be no greater than the drug era end date.
}
