% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelBuildLangFn.R
\name{createDatabaseConnectionLang}
\alias{createDatabaseConnectionLang}
\title{Create details for database connection}
\usage{
createDatabaseConnectionLang(
  connectionDetails = NULL,
  vocabularyDatabaseSchema = NULL
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL which will create dummy credentials}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary format resides. Note that
for SQL Server, this should include both the database and schema
name, for example 'vocabulary.dbo'.}
}
\value{
r language to generate the connection to dbms. Be cautious to not expose credentials
}
\description{
This function will create the database connection in the script
}
