% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateFn.R
\name{createDateOffsetEndStrategy}
\alias{createDateOffsetEndStrategy}
\title{Function creates a date offset end strategy}
\usage{
createDateOffsetEndStrategy(
  offset,
  eventDateOffset = c("StartDate", "EndDate")
)
}
\arguments{
\item{offset}{an integer value specifying padding to the cohort exit.}

\item{eventDateOffset}{an input only for DateOffset specifying whether to add an offset to the start or end of an event
(i.e. StartDate, EndDate)}
}
\value{
This function returns a component class object which contains the end strategy object
}
\description{
This function creates a date offset end strategy. From the ATLAS page: the event end date is derived from adding
a number of days to the event's start or end date. If an offset is added to the event's start date,
all cohort episodes will have the same fixed duration (subject to further censoring). If an offset
is added to the event's end date, persons in the cohort may have varying cohort duration times due to the
varying event durations (such as eras of persistent drug exposure or visit length of stay).
This event persistence assures that the cohort end date will be no greater than the selected index event date,
plus the days offset.
}
