% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userCreateAttributeFn.R
\name{createDeathSourceConceptAttribute}
\alias{createDeathSourceConceptAttribute}
\title{create Death source concept}
\usage{
createDeathSourceConceptAttribute(ConceptSetExpression)
}
\arguments{
\item{ConceptSetExpression}{the concepte set expression we wish to deploy as a source concept attribute
This concept set expression should contain source codes, which may be non-standard.}
}
\value{
a source concept attribute component
}
\description{
create Death source concept
}
