% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userCreateAttributeFn.R
\name{createDeathTypeExcludeAttribute}
\alias{createDeathTypeExcludeAttribute}
\title{create exclude attribute for death type}
\usage{
createDeathTypeExcludeAttribute(logic = FALSE)
}
\arguments{
\item{logic}{toggle FALSE to not exclude}
}
\value{
a component of attribute class
}
\description{
This function creates a attribute for exclusion
}
