% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateAttributeFn.R
\name{createDoseUnitAttribute}
\alias{createDoseUnitAttribute}
\title{create DoseUnit as a concept Attribute}
\usage{
createDoseUnitAttribute(
  conceptIds,
  connectionDetails = NULL,
  connection = NULL,
  vocabularyDatabaseSchema = NULL,
  oracleTempSchema = NULL,
  mapToStandard = TRUE
)
}
\arguments{
\item{conceptIds}{a vector of concept ids. Must be connected to an OMOP vocabulary to use function}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'vocabulary.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
privileges for storing temporary tables.}

\item{mapToStandard}{a logical that indicates whether the concept Ids should be mapped to standard concepts}
}
\value{
a componet of attribute class
}
\description{
This function creates an attribute out of concept values. input concept ids to actionize them within the attribute.
One must clarify if the concept ids should be mapped to standard (default is TRUE) or leave them as is. User needs
to be connected to an OMOP vocabulary to use the lookup functions.
}
