% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userCreateAttributeFn.R
\name{createDrugTypeExcludeAttribute}
\alias{createDrugTypeExcludeAttribute}
\title{create exclude attribute for drug type}
\usage{
createDrugTypeExcludeAttribute(logic = FALSE)
}
\arguments{
\item{logic}{toggle FALSE to not exclude}
}
\value{
a component of attribute class
}
\description{
This function creates a attribute for exclusion
}
