% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateAttributeFn.R
\name{createEffectiveDrugDoseAttribute}
\alias{createEffectiveDrugDoseAttribute}
\title{create EffectiveDrugDose Attribute}
\usage{
createEffectiveDrugDoseAttribute(Op, Value, Extent = NULL)
}
\arguments{
\item{Op}{defines logic for interpreting the numeric or date value.}

\item{Value}{a numeric for the EffectiveDrugDose}

\item{Extent}{a numeric for the EffectiveDrugDose only used if the op is bt or !bt}
}
\value{
a component of attribute class
}
\description{
This function creates an Operator attribute for person EffectiveDrugDose. The user selects the type of
operator, value which is the minimal bound and extent which is the end point of a between bound. Extent is
only used if the op is bt or !bt.
}
