% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateAttributeFn.R
\name{createGapDaysAttribute}
\alias{createGapDaysAttribute}
\title{create GapDays Attribute}
\usage{
createGapDaysAttribute(Op, Value, Extent = NULL)
}
\arguments{
\item{Op}{defines logic for interpreting the numeric or date value.}

\item{Value}{an integer for the GapDays}

\item{Extent}{an integer for the GapDays only used if the op is bt or !bt}
}
\value{
a component of attribute class
}
\description{
This function creates an Operator attribute for person GapDays. The user selects the type of
operator, value which is the minimal bound and extent which is the end point of a between bound. Extent is
only used if the op is bt or !bt.
}
