% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userCreateAttributeFn.R
\name{createGenderAttribute}
\alias{createGenderAttribute}
\title{create gender as a concept Attribute}
\usage{
createGenderAttribute(conceptIds, mapToStandard = TRUE)
}
\arguments{
\item{conceptIds}{a vector of concept ids. Must be connected to an OMOP vocabulary to use function}

\item{mapToStandard}{a logical that indicates whether the concept Ids should be mapped to standard concepts}
}
\value{
a componet of attribute class
}
\description{
This function creates an attribute out of concept values. input concept ids to actionize them within the attribute.
One must clarify if the concept ids should be mapped to standard (default is TRUE) or leave them as is. User needs
to be connected to an OMOP vocabulary to use the lookup functions.
}
