% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateFn.R
\name{createGroup}
\alias{createGroup}
\title{Function creates a group object}
\usage{
createGroup(
  Name,
  type = c("ALL", "ANY", "AT_LEAST", "AT_MOST"),
  count = NULL,
  criteriaList = NULL,
  demographicCriteriaList = NULL,
  Groups = NULL,
  Description = NULL
)
}
\arguments{
\item{Name}{a character string naming the group object, this is required for the object. One should make the name
descriptive of what the group is trying to identify.}

\item{type}{a character string expressing the combination of qualifiying criterias for restriction. Valid options are
ALL meaning all aspects of the group must be true to enter cohort, ANY meaning at least 1 aspect of the group must
be true, AT_LEAST meaning at least a certain count of the group must be true of AT_MOST meaning at most a certain count
must be true of the group. The type entry must be in all capital letters}

\item{count}{the count of criterias needed for restriction. The count only applies if the type if AT_LEAST or AT_MOST.
Otherwise this parameter remains NULL}

\item{criteriaList}{a list of component class count objects to be added. May be left empty, but at least one of
criteriaList, demographicCriteriaList and Groups must be filled. The input must be a list of components}

\item{demographicCriteriaList}{a list of select component class attributes to be added. May be left empty,
but at least one of criteriaList, demographicCriteriaList and Groups must be filled. The input must be a list of components}

\item{Groups}{a list of component class groups to be added. May be left empty,
but at least one of criteriaList, demographicCriteriaList and Groups must be filled. The input must be a list of components}

\item{Description}{a character string describing the count object, this is optional so default is null}
}
\value{
This function returns a component class object which contains the group object and attached concept
set expressions
}
\description{
This function creates a group object of the cohort definition. The group object binds multiple queries, counts,
attributes and other groups to create one component. For entry into the cohort the patient must have a valid instance
of all aspects of the group. Groups are used in additional criteria, inclusion rules and correlated criteria. One can
attach a list of counts as a criteria list, a list of demographic criteria (select attributes) or a list of sub groups.
}
