% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelBuildLangFn.R
\name{createGroupCall}
\alias{createGroupCall}
\title{Get groups from cohort expression and prepare R language}
\usage{
createGroupCall(x, nm, assignName = NULL)
}
\arguments{
\item{x}{the circe cohort definition}

\item{nm}{the naming convention for sub-objects}

\item{assignName}{the naming convention to assign the object}
}
\value{
r language to generate the groups of the cohort
}
\description{
This function creates groups from cohort and turns them into
R language which will then create them as a CAPR objects
}
