% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelCreateFn.R
\name{createLogicalAttribute}
\alias{createLogicalAttribute}
\title{createLogicalAttribue}
\usage{
createLogicalAttribute(name, logic = TRUE)
}
\arguments{
\item{name}{is the name of the attribute}

\item{logic}{whether the logic is true or false, default is true}
}
\description{
createLogicalAttribue
}
