% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateAttributeFn.R
\name{createMeasurementSourceConceptAttribute}
\alias{createMeasurementSourceConceptAttribute}
\title{create measurement source concept}
\usage{
createMeasurementSourceConceptAttribute(ConceptSetExpression)
}
\arguments{
\item{ConceptSetExpression}{the concept set expression we wish to deploy as a source concept attribute
This concept set expression should contain source codes, which may be non-standard.}
}
\value{
a source concept attribute component
}
\description{
create measurement source concept
}
