% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateAttributeFn.R
\name{createObservationTypeExcludeAttribute}
\alias{createObservationTypeExcludeAttribute}
\title{create exclude attribute for observation type}
\usage{
createObservationTypeExcludeAttribute(logic = FALSE)
}
\arguments{
\item{logic}{toggle FALSE to not exclude}
}
\value{
a component of attribute class
}
\description{
This function creates a attribute for exclusion
}
