% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userCreateFn.R
\name{createObservationWindow}
\alias{createObservationWindow}
\title{Function creates an Observation Window}
\usage{
createObservationWindow(PriorDays = 0L, PostDays = 0L)
}
\arguments{
\item{PriorDays}{number of days prior to the initial event of continuous observation}

\item{PostDays}{number of days of continous observation after index date}
}
\value{
This function returns a observation window class object providing prior and post days of observation
}
\description{
This function creates an observation window used in a primary criteria. The observation window provides the amount of
time before and after the initial event of continuous observation necessary for a person to be eligible to enter the cohort.
The minimal observation days would be 0 days of prior observation and 0 days of post observations. This is the default for
this function.
}
