% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelCreateFn.R
\name{createOpAttribute}
\alias{createOpAttribute}
\title{createOpAttribute}
\usage{
createOpAttribute(Name, Op, Value, Extent = NULL)
}
\arguments{
\item{Name}{a name}

\item{Op}{a type of operator}

\item{Value}{a value either integer or character for dates}

\item{Extent}{only if Op is bt or !bt, otherwise NULL. Value is either integer or character for dates}
}
\description{
createOpAttribute
}
