% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateAttributeFn.R
\name{createPeriodStartDateAttribute}
\alias{createPeriodStartDateAttribute}
\title{create period Start Date Attribute}
\usage{
createPeriodStartDateAttribute(Op, Value, Extent = NULL)
}
\arguments{
\item{Op}{defines logic for interpreting the numeric or date value.}

\item{Value}{a character string of the date}

\item{Extent}{a character string of the extent only used if the op is bt or !bt}
}
\value{
a componet of attribute class
}
\description{
This function creates an Operator attribute for the period start date. The user selects the type of
operator, value which is the minimal bound and extent which is the end point of a between bound. Extent is
only used if the op is bt or !bt.
}
