% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateFn.R
\name{createPrimaryCriteria}
\alias{createPrimaryCriteria}
\title{Function creates a Primary Criteria}
\usage{
createPrimaryCriteria(
  Name,
  ComponentList,
  ObservationWindow = NULL,
  Limit,
  Description = NULL
)
}
\arguments{
\item{Name}{a character string naming the group object, this is required for the object. One should make the name
descriptive of what the group is trying to identify.}

\item{ComponentList}{a list of query components to add to the primary criteria. These components include
the queries and concept set expression used in the cohort.}

\item{ObservationWindow}{an observationWindow class object that set the prior and post days of continuous
observation for the initial event}

\item{Limit}{how to limit initial events per person}

\item{Description}{a character string describing the count object, this is optional so default is null}
}
\value{
new primary criteria component.
}
\description{
Function creates a primary criteria from multiple queries. User adds a list of component class queries, identifies the
observation window and the criteria limit.
}
