% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelCreateFn.R
\name{createQuery}
\alias{createQuery}
\title{createQuery}
\usage{
createQuery(
  Domain,
  Component = NULL,
  attributeList = NULL,
  Name = NULL,
  Description = NULL
)
}
\arguments{
\item{Domain}{list the domain from the table we are searching in the query}

\item{Component}{add the concept set expression we want to query}

\item{attributeList}{a list of attribute class components to add, if not attributes keep null}

\item{Name}{is the name of query, optional}

\item{Description}{an optional description of the query}
}
\description{
createQuery
}
