% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateAttributeFn.R
\name{createRangeHighAttribute}
\alias{createRangeHighAttribute}
\title{create RangeHigh Attribute}
\usage{
createRangeHighAttribute(Op, Value, Extent = NULL)
}
\arguments{
\item{Op}{defines logic for interpreting the numeric or date value.}

\item{Value}{an integer for the RangeHigh}

\item{Extent}{an integer for the RangeHigh only used if the op is bt or !bt}
}
\value{
a component of attribute class
}
\description{
This function creates an Operator attribute for person RangeHigh. The user selects the type of
operator, value which is the minimal bound and extent which is the end point of a between bound. Extent is
only used if the op is bt or !bt.
}
