% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateAttributeFn.R
\name{createRouteConceptsAttribute}
\alias{createRouteConceptsAttribute}
\title{create RouteConcepts as a concept Attribute}
\usage{
createRouteConceptsAttribute(
  conceptIds,
  connectionDetails = NULL,
  connection = NULL,
  vocabularyDatabaseSchema = NULL,
  tempEmulationSchema = NULL,
  mapToStandard = TRUE
)
}
\arguments{
\item{conceptIds}{a vector of concept ids. Must be connected to an OMOP vocabulary to use function}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary format resides. Note that
for SQL Server, this should include both the database and schema
name, for example 'vocabulary.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}

\item{mapToStandard}{a logical that indicates whether the concept Ids should be mapped to standard concepts}
}
\value{
a componet of attribute class
}
\description{
This function creates an attribute out of concept values. input concept ids to actionize them within the attribute.
One must clarify if the concept ids should be mapped to standard (default is TRUE) or leave them as is. User needs
to be connected to an OMOP vocabulary to use the lookup functions.
}
