% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelCreateFn.R
\name{createSourceConceptAttribute}
\alias{createSourceConceptAttribute}
\title{createSourceConceptAttribute}
\usage{
createSourceConceptAttribute(Domain, ConceptSetExpression)
}
\arguments{
\item{Domain}{the type of domain for the source concept}

\item{ConceptSetExpression}{the concept set expression component to add}
}
\description{
createSourceConceptAttribute
}
