% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userCreateFn.R
\name{createTimeline}
\alias{createTimeline}
\title{Set the Timeline in the criteria}
\usage{
createTimeline(
  StartWindow,
  EndWindow = NULL,
  RestrictVisit = FALSE,
  IgnoreObservationPeriod = FALSE
)
}
\arguments{
\item{StartWindow}{a window class object that modifies when to begin monitoring for an observation}

\item{EndWindow}{a window class object that ends the time observing events. This window is not always created so the
default is NULL, initializing an empty window}

\item{RestrictVisit}{a logic toggle where TRUE restricts to the same visit}

\item{IgnoreObservationPeriod}{a logic toggle where TRUE allows events outside the observation period}
}
\value{
a new Timeline class object
}
\description{
When a criteria object is initialized a default timeline object is also initialized. To change the timeline object
we set it to a new information. Inputs include StartWindow, EndWindow, RestrictVisit, and IgnoreObservationPeriod.
The StartWindow and EndWindow inputs require a window class object. A new window can be initialized using the
createWindow function.
}
