% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelBuildLangFn.R
\name{createTimelineCall}
\alias{createTimelineCall}
\title{Function to create a timeline call}
\usage{
createTimelineCall(x, objectName)
}
\arguments{
\item{x}{the circe cohort definition}

\item{objectName}{the naming convention to assign the object}
}
\value{
r language to generate the timelines of the cohort
}
\description{
Function to create a timeline call
}
