% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateAttributeFn.R
\name{createVisitSourceConceptAttribute}
\alias{createVisitSourceConceptAttribute}
\title{create Visit source concept}
\usage{
createVisitSourceConceptAttribute(ConceptSetExpression)
}
\arguments{
\item{ConceptSetExpression}{the concept set expression we wish to deploy as a source concept attribute
This concept set expression should contain source codes, which may be non-standard.}
}
\value{
a source concept attribute component
}
\description{
create Visit source concept
}
