% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userCreateAttributeFn.R
\name{createVisitTypeExcludeAttribute}
\alias{createVisitTypeExcludeAttribute}
\title{create exclude attribute for visit type}
\usage{
createVisitTypeExcludeAttribute(logic = FALSE)
}
\arguments{
\item{logic}{toggle FALSE to not exclude}
}
\value{
a component of attribute class
}
\description{
This function creates a attribute for exclusion
}
