% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userCreateFn.R
\name{createWindow}
\alias{createWindow}
\title{Function to initialize a new window object}
\usage{
createWindow(
  StartDays,
  StartCoeff = c("Before", "After"),
  EndDays,
  EndCoeff = c("Before", "After"),
  EventStarts = TRUE,
  IndexStart = TRUE
)
}
\arguments{
\item{StartDays}{number of days at start of window}

\item{StartCoeff}{where to begin counting relative to index date: before or after}

\item{EndDays}{number of days to end window}

\item{EndCoeff}{where to end counting relative to index date: before or after}

\item{EventStarts}{if TRUE then this counts from the start of an event otherwise from the end of an event}

\item{IndexStart}{if TRUE then the index date is the start of event otherwise the end of an event}
}
\value{
a new window class object
}
\description{
A window depicts the timeline from which events are counted. The window has four components: Start, End, EventStart,
and Index Start. First, we determine whether observations are viewed from the start of the event or at the end.
By default EventStart is TRUE. Next the start of recording is identied using days and coefficient. The coefficient
distinguishes how the days are counted relative to the index date. The end recording is the same as the start,
now identifying the end of observation. Finally it is identified whether the index date is relative the start or
end of occurrence. A timeline has a start and end window. Usually the end window is not defined. An End Window
adds a constraint to the Start Window of a timeline
}
