% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelBuildLangFn.R
\name{createWindowCall}
\alias{createWindowCall}
\title{Function to create a window object call}
\usage{
createWindowCall(x)
}
\arguments{
\item{x}{the circe cohort definition}
}
\value{
r language to generate the windows of the cohort
}
\description{
Function to create a window object call
}
