% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\docType{class}
\name{Criteria-class}
\alias{Criteria-class}
\title{An S4 for a criteria}
\description{
a criteria is a temporal observation of a clinical event relative to the index event
}
\section{Slots}{

\describe{
\item{\code{occurrence}}{an occurrence object specifying how many events must occur
to consider the event as part of the cohort definition}

\item{\code{query}}{a query object that provides context to the clinical event of interest}

\item{\code{aperture}}{an eventAperture object that shows the temporal span where the event is to be observed
relative to the index event}
}}

