% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{criteria}
\alias{criteria}
\title{Function to construct a criteria object}
\usage{
criteria(occurrence, query, aperture)
}
\arguments{
\item{occurrence}{an occurrence object specifying how many events must occur
to consider the event as part of the cohort definition}

\item{query}{a query object that provides context to the clinical event of interest}

\item{aperture}{an eventAperture object that shows the temporal span where the event is to be observed
relative to the index event}
}
\description{
Function to construct a criteria object
}
