% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSet.R
\name{cs}
\alias{cs}
\alias{exclude}
\alias{mapped}
\alias{descendants}
\title{Create a concept set}
\usage{
cs(..., name)

exclude(...)

mapped(...)

descendants(...)
}
\arguments{
\item{...}{One or more numeric vectors that can be coerced to integers, or
Calls to helper functions "exclude", "descendants", or "mapped".}

\item{name}{A name for the concept set}
}
\value{
A Capr Concept Set Object

A list of Capr concepts

A list of Capr concepts

A list of Capr concepts
}
\description{
cs is used to create concept set expressions.

`exclude` is meant to be used inside `cs` when creating a new concept set.

`mapped` is meant to be used inside `cs` when creating a new concept set.

`descendants` is meant to be used inside `cs` when creating a new concept set.
}
\section{Functions}{
\itemize{
\item \code{exclude()}: exclude concepts

\item \code{mapped()}: Include mapped concepts

\item \code{descendants()}: Include descendants

}}
\examples{
cs(1, 2, name = "concepts")
cs(1, c(10, 11, 2), name = "concepts")
cs(1, seq(2, 10, 2), name = "concepts")
cs(1, 2, 3, exclude(4, 5), name = "concepts")
cs(1, 2, 3, exclude(4, 5), mapped(6, 7), name = "concepts")
cs(1, 2, 3, exclude(4, 5), mapped(6, 7), descendants(8, 9), name = "concepts")
cs(descendants(1, 2, 3),  exclude(descendants(8, 9)), name = "concepts")
}
