% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelClasses.R
\docType{class}
\name{CustomEraEndStrategy-class}
\alias{CustomEraEndStrategy-class}
\title{An S4 class for CustomEraEndStrategy}
\description{
An end strategy class specifying the time until the end of drug use for cohort exit
}
\section{Slots}{

\describe{
\item{\code{DrugCodesetId}}{the guid of the drug concept set expression to activate in the end strategy}

\item{\code{GapDays}}{an integer showing the maximum allowable days between successive exposures.}

\item{\code{Offset}}{an integer value specifying padding to the cohort exit.}
}}

