% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-dateAdjustment.R
\docType{class}
\name{dateAdjustmentAttribute-class}
\alias{dateAdjustmentAttribute-class}
\title{An S4 class for a date adjustment attribute}
\description{
An S4 class for a date adjustment attribute
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the attribute}

\item{\code{startWith}}{character string either START_DATE or END_DATE}

\item{\code{startOffset}}{an integer value, default 0}

\item{\code{endWith}}{character string either START_DATE or END_DATE}

\item{\code{endOffset}}{an integer value, default 0}
}}

