% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\docType{class}
\name{DateOffsetEndStrategy-class}
\alias{DateOffsetEndStrategy-class}
\title{An S4 class for DateOffsetEndStrategy}
\description{
An end strategy class specifying a number of days from the start or end of the initial event until
cohort exit
}
\section{Slots}{

\describe{
\item{\code{DateField}}{a character string specifying either the StartDate or EndDate of the initial event to begin counting
days until cohort exit}

\item{\code{Offset}}{an integer value specifying padding to the cohort exit.}
}}

