% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSet.R
\name{descendants}
\alias{descendants}
\title{Include descendants in a concept set}
\usage{
descendants(...)
}
\arguments{
\item{...}{One or more numeric vectors that can be coerced to integers, or
Calls to helper functions "exclude", "descendants", or "mapped".}
}
\value{
A list of Capr concepts
}
\description{
`descendants` is meant to be used inside `cs` when creating a new concept set.
}
