% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{deviceExposure}
\alias{deviceExposure}
\title{Query the drug domain}
\usage{
deviceExposure(conceptSet, ...)
}
\arguments{
\item{conceptSet}{A drug concept set (optional)}

\item{...}{optional attributes}
}
\value{
A Capr Query
}
\description{
Query the drug domain
}
