% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-logic.R
\name{deviceTypeExclude}
\alias{deviceTypeExclude}
\title{DeviceTypeExclude query attribute}
\usage{
deviceTypeExclude(exclude = FALSE)
}
\arguments{
\item{exclude}{logical; FALSE = include all types (default), TRUE = exclude}
}
\value{
An attribute for use in \code{deviceExposure()}
}
\description{
DeviceTypeExclude query attribute
}
