% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{doseEra}
\alias{doseEra}
\title{Query the dose era domain}
\usage{
doseEra(conceptSet, ...)
}
\arguments{
\item{conceptSet}{A drug ingredient concept set (optional)}

\item{...}{optional attributes}
}
\value{
A Capr Query
}
\description{
Query the dose era domain
}
