% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-conceptSetSelection.R
\name{doseUnitCS}
\alias{doseUnitCS}
\title{Dose unit filter via concept set — DrugExposure}
\usage{
doseUnitCS(conceptSet, isExclusion = FALSE)
}
\arguments{
\item{conceptSet}{A \code{ConceptSet} built with \code{cs()}.}

\item{isExclusion}{logical; if \code{TRUE} the types in the concept set are excluded.}
}
\value{
An attribute for use in \code{drugExposure()}.
}
\description{
Dose unit filter via concept set — DrugExposure
}
