% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{doseValue}
\alias{doseValue}
\title{Dose value attribute for DoseEra}
\usage{
doseValue(op)
}
\arguments{
\item{op}{an opAttribute object (numeric) that defines the logical operation and value}
}
\value{
An attribute for use in \code{\link{doseEra}()}
}
\description{
Filter dose era criteria by dose value (e.g. \code{gt(92)} for dose > 92).
Used only in a dose era query.
}
