% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit.R
\name{drugExit}
\alias{drugExit}
\title{Function to create an exit based on exit based on the end of a continuous drug exposure}
\usage{
drugExit(
  conceptSet,
  persistenceWindow = 0L,
  surveillanceWindow = 0L,
  daysSupplyOverride = NULL
)
}
\arguments{
\item{conceptSet}{the concept set of the drug exposure used to identify the exit}

\item{persistenceWindow}{allow for a maximum of days between exposure records when inferring the era
of persistence exposure}

\item{surveillanceWindow}{add days to the end of the era of persistence exposure as an additional period of
surveillance prior to cohort exit}

\item{daysSupplyOverride}{force drug exposure days supply to a set number of days}
}
\description{
Function to create an exit based on exit based on the end of a continuous drug exposure
}
