% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit.R
\docType{class}
\name{DrugExposureExit-class}
\alias{DrugExposureExit-class}
\title{An S4 class for a cohort exit based on continuous exposure persistence.}
\description{
Specify a concept set that contains one or more drugs.
A drug era will be derived from all drug exposure events for any of the drugs within the specified concept set,
using the specified persistence window as a maximum allowable gap in days between successive exposure
events and adding a specified surveillance window to the final exposure event.
If no exposure event end date is provided, then an exposure event end date is inferred to
be event start date + days supply in cases when days supply is available or
event start date + 1 day otherwise. This event persistence assures that the cohort end date
will be no greater than the drug era end date.
}
\section{Slots}{

\describe{
\item{\code{conceptSet}}{the concept set of the drug exposure used to identify the exit}

\item{\code{persistenceWindow}}{allow for a maximum of days between exposure records when inferring the era
of persistence exposure}

\item{\code{surveillanceWindow}}{add days to the end of the era of persistence exposure as an additional period of
surveillance prior to cohort exit}

\item{\code{daysSupplyOverride}}{force drug exposure days supply to a set number of days}

\item{\code{count}}{an integer specifying the number of occurrences for a criteria}
}}

