% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{drugSourceConcept}
\alias{drugSourceConcept}
\title{Add a drug source concept attribute}
\usage{
drugSourceConcept(conceptSet)
}
\arguments{
\item{conceptSet}{a ConceptSet object containing the source concepts}
}
\value{
An attribute that can be used in a query function
}
\description{
Add a drug source concept attribute
}
