% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-logic.R
\name{drugTypeExclude}
\alias{drugTypeExclude}
\title{DrugTypeExclude query attribute}
\usage{
drugTypeExclude(exclude = FALSE)
}
\arguments{
\item{exclude}{logical; FALSE = include all types (default), TRUE = exclude}
}
\value{
An attribute for use in \code{drugExposure()}
}
\description{
DrugTypeExclude query attribute
}
