% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserEditFn.R
\name{editConceptSetItem}
\alias{editConceptSetItem}
\title{Function to edit a concept set item}
\usage{
editConceptSetItem(obj, edit, index = NULL, mapping = NULL, newName = NULL)
}
\arguments{
\item{obj}{the component you wish to edit}

\item{edit}{the edit to make}

\item{index}{an index to specify a postion in a list}

\item{mapping}{a character of includeDescendants, isExcluded or includeMapped to toggle logic}

\item{newName}{a character string updating the name of the concept set expression}
}
\value{
the edit s4 class object
}
\description{
This function edits a concept set item class
}
