% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserEditFn.R
\name{editCount}
\alias{editCount}
\title{Function to edit Meta data}
\usage{
editCount(obj, edit, slotNms)
}
\arguments{
\item{obj}{the component you wish to edit}

\item{edit}{the edit to make}

\item{slotNms}{a list object where each entry is a slot across multiple objects. The list must
be constructed in order and can be done by following the object structure. For example to edit
a window in the timeline one must construct a list('Timeline', 'StartWindow','Start'). If one
wants to edit the count in an occurrence the list is: list('Occurrence','Count').}
}
\value{
the edit s4 class object
}
\description{
This function edits a meta data class
}
