% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserEditFn.R
\name{editExpressionType}
\alias{editExpressionType}
\title{Function to edit Expression type}
\usage{
editExpressionType(obj, edit, slotNm)
}
\arguments{
\item{obj}{the component you wish to edit}

\item{edit}{the edit to make}

\item{slotNm}{the slot to edit}
}
\value{
the edit s4 class object
}
\description{
This function edits a expression type class
}
