% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserEditFn.R
\name{editInclusionRules}
\alias{editInclusionRules}
\title{Function to edit Inclusion Rules}
\usage{
editInclusionRules(
  inclusionRules,
  edit,
  detail = c("Name", "Description", "Rule", "Limit"),
  add = FALSE,
  index = NULL
)
}
\arguments{
\item{inclusionRules}{the inclusion rules component you wish to edit}

\item{edit}{the edit to make. The edit must conform to the structure of the location detail
where the edit is made. See detail for more information}

\item{detail}{the slot to edit in the inclusion rules. Options are: Name, Description, Rule and Limit.
If editing the name or description the edit must be a character string. If editing the limit
the edit must be a character string of either All, First or Last. If the detail is a rule,
the edit must be a Group type component class. One can use the function componentType to
check the type for a component class object.}

\item{add}{a loggic toggle to say if you are adding a piece to the pc component}

\item{index}{an index to specify the position in a list that is to be modified. If null defaults to 1}
}
\value{
the edit s4 class object
}
\description{
This function edits a meta data class
}
