% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserEditFn.R
\name{editLimit}
\alias{editLimit}
\title{Function to edit Limit}
\usage{
editLimit(obj, edit = c("All", "First", "Last"))
}
\arguments{
\item{obj}{the component you wish to edit}

\item{edit}{the edit to make either all first or last}
}
\value{
the edited s4 class object
}
\description{
This function edits a limit class
}
