% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserEditFn.R
\name{editObservationWindow}
\alias{editObservationWindow}
\title{Function to edit Observation Window}
\usage{
editObservationWindow(obj, slotNm, edit)
}
\arguments{
\item{obj}{the component you wish to edit}

\item{slotNm}{the slot to edit}

\item{edit}{the edit to make}
}
\value{
the edit s4 class object
}
\description{
This function edits a observation window class
}
