% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserEditFn.R
\name{editQuery}
\alias{editQuery}
\title{Function to edit Query}
\usage{
editQuery(obj, edit, slotNm, index = NULL)
}
\arguments{
\item{obj}{the component you wish to edit}

\item{edit}{the edit to make}

\item{slotNm}{the slot to edit}

\item{index}{an integer index specifying the location within a list, if not needed leave null}
}
\value{
the edit s4 class object
}
\description{
This function edits a query class
}
