% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserEditFn.R
\name{editTimeline}
\alias{editTimeline}
\title{Function to edit Timeline}
\usage{
editTimeline(obj, slotNm, edit)
}
\arguments{
\item{obj}{the component you wish to edit}

\item{slotNm}{the slot to edit}

\item{edit}{the edit to make}
}
\value{
the edit s4 class object
}
\description{
This function edits a timeline class
}
