% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{endDate}
\alias{endDate}
\title{Function that creates a end date attribute}
\usage{
endDate(op, type = "occurrence")
}
\arguments{
\item{op}{an opAttribute object must be a date that defines the logical operation used to
determine eligible end dates}

\item{type}{specify the type of date to use either occurrence or era. default as occurrence}
}
\value{
An attribute that can be used in a cohort definition
}
\description{
Function that creates a end date attribute
}
