% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelClasses.R
\docType{class}
\name{EndOfCtsObsEndStrategy-class}
\alias{EndOfCtsObsEndStrategy-class}
\title{An S4 class for EndOfCtsObsEndStrategy}
\description{
When the end strategy is not defined the cohort exit is done based on the end of continuous observation.
This class is an end strategy type.
}
\section{Slots}{

\describe{
\item{\code{EndOfContinuousObservation}}{set as true for end strategy option}
}}

