% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\docType{class}
\name{Endpoint-class}
\alias{Endpoint-class}
\title{An S4 class for an Endpoint}
\description{
this determines the time in days relative to an index either before or after
}
\section{Slots}{

\describe{
\item{\code{days}}{either a character string all or an integer for the number of days}

\item{\code{coeff}}{a character string either before or after}
}}

