% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{endsWith}
\alias{endsWith}
\alias{endsWith,character-method}
\title{String ends with operator}
\usage{
endsWith(x)

\S4method{endsWith}{character}(x)
}
\arguments{
\item{x}{the suffix string to search for}
}
\description{
function that builds an opAttribute based on string suffix logic
}
