% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{eq}
\alias{eq}
\alias{eq,integer-method}
\alias{eq,numeric-method}
\alias{eq,Date-method}
\alias{eq,character-method}
\title{Equal to operator}
\usage{
eq(x)

\S4method{eq}{integer}(x)

\S4method{eq}{numeric}(x)

\S4method{eq}{Date}(x)

\S4method{eq}{character}(x)
}
\arguments{
\item{x}{the value to used as a bound in the op logic. This can either be an integer, numeric, or
Date data type. Different data types will return the appropriate opAttribute type}
}
\description{
function that builds an opAttribute based on equal to logic
}
