% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSet.R
\docType{methods}
\name{==}
\alias{==}
\alias{==,ConceptSet,ConceptSet-method}
\title{FUnction checks if two concept set class objects are equivalent}
\usage{
\S4method{==}{ConceptSet,ConceptSet}(e1, e2)
}
\arguments{
\item{e1, e2}{a ConceptSet Class object}
}
\description{
FUnction checks if two concept set class objects are equivalent
}
