% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{eraLength}
\alias{eraLength}
\title{Era length attribute for DrugEra}
\usage{
eraLength(op)
}
\arguments{
\item{op}{an opAttribute object (integer) that defines the logical operation and value}
}
\value{
An attribute for use in \code{\link{drugEra}()}
}
\description{
Filter drug era criteria by era length in days (e.g. \code{gt(14L)} for era longer than 14 days).
Used only in a drug era query.
}
