% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\docType{class}
\name{EventAperture-class}
\alias{EventAperture-class}
\title{An S4 class for Aperture}
\description{
The aperture class provides context to when the criteria must be observed in a person timeline to pretain to the expression
}
\section{Slots}{

\describe{
\item{\code{startWindow}}{a EventWindow class object identifying the start window}

\item{\code{endWindow}}{a EventWindow class object ifentifying the end window (optional)}

\item{\code{restrictVisit}}{a logic toggle where TRUE restricts to the same visit}

\item{\code{ignoreObservationPeriod}}{a logic toggle where TRUE allows events outside the observation period}
}}

