% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{eventStarts}
\alias{eventStarts}
\title{Function creates an event window where the event starts}
\usage{
eventStarts(a, b, index = c("startDate", "endDate"))
}
\arguments{
\item{a}{the left side of the event window}

\item{b}{the right side of the event window}

\item{index}{specifying what part of the index we start looking for events
either at the index start date or index enddate}
}
\description{
Function creates an event window where the event starts
}
