% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{exact}
\alias{exact}
\alias{exact,character-method}
\title{String exact match operator}
\usage{
exact(x)

\S4method{exact}{character}(x)
}
\arguments{
\item{x}{the exact string to match}
}
\description{
function that builds an opAttribute based on exact string match logic
}
