% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSet.R
\name{exclude}
\alias{exclude}
\title{Exclude concepts from a concept set}
\usage{
exclude(...)
}
\arguments{
\item{...}{One or more numeric vectors that can be coerced to integers, or
Calls to helper functions "exclude", "descendants", or "mapped".}
}
\value{
A list of Capr concepts
}
\description{
`exclude` is meant to be used inside `cs` when creating a new concept set.
}
