% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{exit}
\alias{exit}
\title{Define cohort exit strategy}
\usage{
exit(endStrategy, censor = NULL)
}
\arguments{
\item{endStrategy}{An end strategy object: \code{\link{observationExit}},
\code{\link{fixedExit}}, or \code{\link{drugExit}}.}

\item{censor}{Optional \code{\link{censoringEvents}} object of queries that end the cohort
early if they occur after the index date.}
}
\value{
A \code{CohortExit} object.
}
\description{
Define cohort exit strategy
}
\examples{
exit(endStrategy = observationExit())
exit(endStrategy = fixedExit(offsetDays = 30L))
}
\seealso{
\code{\link{observationExit}}, \code{\link{fixedExit}}, \code{\link{drugExit}},
  \code{\link{censoringEvents}}, \code{\link{cohort}}
}
