% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\docType{class}
\name{ExpressionType-class}
\alias{ExpressionType-class}
\title{An S4 class for Expression type}
\description{
An expression type quantifies the number of criteria's needed to set as restriction. Types include:
All, Any, at least and at most. If the expression type is at least or at most a count is required
to express the type
}
\section{Slots}{

\describe{
\item{\code{Type}}{boolean operator for the number of items in group to include. all, any, at most and at least}

\item{\code{Count}}{the number of criteria's needed for restriction. If Type is ALL or ANY this value is NA}
}}

