% Generated by roxygen2 (4.0.2): do not edit by hand
\name{file_rc}
\alias{file_rc}
\title{Import or Deleta a file attached to a REDCap record}
\usage{
file_rc(url, certificate, token, action, id, field, file = NULL,
  event = NULL, verbose = FALSE)
}
\arguments{
\item{url}{The url for your REDCap project.}

\item{certificate}{Certificate file (with path) containing REDCap security}

\item{token}{API token (character string).}

\item{action}{Either import or delete}

\item{id}{Record id to import file to}

\item{field}{Field that holds the file}

\item{file}{Your file (character string--don't forget the path)}

\item{event}{The event you want the file sent to}

\item{verbose}{Logical (default = FALSE). Verbose output from the posting}
}
\description{
Talk to REDCap and either import or deleta a file attached
to a specific record/field. You can only access record/field's file
at a time.
}

