% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-logic.R
\name{firstOccurrence}
\alias{firstOccurrence}
\title{Add first occurrence attribute}
\usage{
firstOccurrence()
}
\value{
An attribute that can be used in a query function
}
\description{
Add first occurrence attribute
}
