% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit.R
\docType{class}
\name{FixedDurationExit-class}
\alias{FixedDurationExit-class}
\title{An S4 class for a cohort exit based on fixed duration persistence.}
\description{
The event end date is derived from adding a number of days to
the event's start or end date. If an offset is added to the event's start date,
all cohort episodes will have the same fixed duration (subject to further censoring).
If an offset is added to the event's end date, persons in the cohort may have
varying cohort duration times due to the varying event durations
(such as eras of persistent drug exposure or visit length of stay).
This event persistence assures that the cohort end date will be no greater
than the selected index event date, plus the days offset.
}
\section{Slots}{

\describe{
\item{\code{index}}{specification of event date to offset}

\item{\code{offsetDays}}{an integer specifying the number of days to offset from the event date}
}}

