% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit.R
\name{fixedExit}
\alias{fixedExit}
\title{Function to create an exit based on exit based on the end of a continuous drug exposure}
\usage{
fixedExit(index = c("startDate", "endDate"), offsetDays)
}
\arguments{
\item{index}{specification of event date to offset. Can be either startDate or endDate}

\item{offsetDays}{an number specifying the days to offset from the event date. Will coerce to an integer}
}
\value{
a fixed Duration exit S4 object used to define the cohort exit as the end of a specified time
}
\description{
Function to create an exit based on exit based on the end of a continuous drug exposure
}
