% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelUtilityFn.R
\name{formatConceptTable}
\alias{formatConceptTable}
\title{Format Concept Table}
\usage{
formatConceptTable(concepts_df)
}
\arguments{
\item{concepts_df}{concept data frame to format}
}
\description{
This is an internal funcion to get concept to circe format
}
