% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/friendify.R
\name{friendify}
\alias{friendify}
\title{Convert REDCap API exports into a friendly data frame}
\usage{
friendify(rc_raw)
}
\arguments{
\item{rc_raw}{Raw data from the export}
}
\value{
Friendly data frame containing your REDCap export
}
\description{
Creates a text connection to the raw REDCAP export and uses \code{read.table}
to make a friendly data frame.
}

