% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{generateCaprTemplate}
\alias{generateCaprTemplate}
\title{Generate a Capr cohort using a template}
\usage{
generateCaprTemplate(file, .capr)
}
\arguments{
\item{file}{the input file of a concept set}

\item{.capr}{a function that creates a capr cohort}
}
\description{
Generate a Capr cohort using a template
}
