% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelBuildLangFn.R
\name{getACCall}
\alias{getACCall}
\title{Get additional criteria from cohort expression and prepare R language}
\usage{
getACCall(x)
}
\arguments{
\item{x}{the circe cohort definition}
}
\value{
r language to generate the additional criteria of the cohort
}
\description{
Get additional criteria from cohort expression and prepare R language
}
