% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelBuildLangFn.R
\name{getCenCall}
\alias{getCenCall}
\title{Get censoring criteria from cohort expression and prepare R language}
\usage{
getCenCall(x)
}
\arguments{
\item{x}{the circe cohort definition}
}
\value{
r language to generate the censoring criteria of the cohort
}
\description{
Get censoring criteria from cohort expression and prepare R language
}
