% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelBuildLangFn.R
\name{getCohortDefinitionCall}
\alias{getCohortDefinitionCall}
\title{Get call to build cohort definition}
\usage{
getCohortDefinitionCall(x, nm = NULL)
}
\arguments{
\item{x}{the circe cohort definition}

\item{nm}{the naming convention to assign the object}
}
\value{
r language to generate the cohort
}
\description{
This function generates the cohort definition call and the
R language calls needed to build the lower level objects for the cohort definition
}
