% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserConceptLookupFn.R
\name{getConceptCodeDetails}
\alias{getConceptCodeDetails}
\title{Lookup Concepts by OMOP Concept Code using Vocabulary}
\usage{
getConceptCodeDetails(
  conceptCode,
  vocabulary,
  connectionDetails = NULL,
  connection = NULL,
  vocabularyDatabaseSchema = NULL,
  oracleTempSchema = NULL,
  mapToStandard = TRUE
)
}
\arguments{
\item{conceptCode}{a character vector of concept codes}

\item{vocabulary}{a single character string with the vocabulary of the codes}

\item{connection}{Either an object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package or or object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. If  \code{connectionDetails}
is provided a new connection will be opened at the start
of the function and closed when the function finishes.}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary format resides. Note that
for SQL Server, this should include both the database and schema
name, for example 'vocabulary.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
privileges for storing temporary tables.}

\item{mapToStandard}{logic to map to standard OMOP concept}
}
\value{
a tibble data frame object with conceptId, conceptName, standardConcept,
              standardConceptCaption, invalidReason, invalidReasonCaption, conceptCode,
              domainId, vocabularyId, conceptClassId.
}
\description{
This function looks up concepts using the OMOP concept code and vocabulary. Function requires a dbms connection to use
}
