% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSearch.R
\name{getConceptDescendants}
\alias{getConceptDescendants}
\title{Get all descendants of one or more standard concepts}
\usage{
getConceptDescendants(
  conceptIds,
  connection,
  vocabularyDatabaseSchema,
  minLevels = 0L,
  maxLevels = Inf
)
}
\arguments{
\item{conceptIds}{Integer vector of standard concept IDs.}

\item{connection}{A DatabaseConnector connection to an OMOP CDM.}

\item{vocabularyDatabaseSchema}{Schema containing the OMOP vocabulary tables. Required.}

\item{minLevels}{Minimum levels of separation (default 0 includes seeds).}

\item{maxLevels}{Maximum levels of separation (default Inf = all descendants).}
}
\value{
A tibble with concept details plus min/max_levels_of_separation.
}
\description{
Traverses concept_ancestor. SQL lives in inst/sql/getConceptDescendants.sql.
}
\examples{
\dontrun{
connection <- DatabaseConnector::connect(Eunomia::getEunomiaConnectionDetails())
getConceptDescendants(313217L, connection, vocabularyDatabaseSchema = "main")
}
}
\seealso{
searchConcepts, descendants, cs
}
