% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelBuildLangFn.R
\name{getConceptSetCall}
\alias{getConceptSetCall}
\title{Get concept sets from cohort expression and prepare R language}
\usage{
getConceptSetCall(x)
}
\arguments{
\item{x}{the circe cohort definition}
}
\value{
r language to generate the concept set expressions of the cohort
}
\description{
This function takes the concept sets from the circe cohort definition
and generates R functions to create them in the R environment. The data
saved is R language to generate the objects. They are evaluated separately
}
