% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelUtilityFn.R
\name{getConceptSetExpression,Component-method}
\alias{getConceptSetExpression,Component-method}
\alias{getConceptSetExpression}
\title{Function to get Concept Set Expressions}
\usage{
\S4method{getConceptSetExpression}{Component}(x)
}
\arguments{
\item{x}{the component to check}
}
\value{
a list of concept set expressions used in the object
}
\description{
Function to get Concept Set Expressions
}
