% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelBuildLangFn.R
\name{getESCall}
\alias{getESCall}
\title{Get end strategy from cohort expression and prepare R language}
\usage{
getESCall(x)
}
\arguments{
\item{x}{the circe cohort definition}
}
\value{
r language to generate the end strategy of the cohort
}
\description{
Get end strategy from cohort expression and prepare R language
}
