% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelBuildLangFn.R
\name{getIRSCall}
\alias{getIRSCall}
\title{Get inclusion rules from cohort expression and prepare R language}
\usage{
getIRSCall(x)
}
\arguments{
\item{x}{the circe cohort definition}
}
\value{
r language to generate the inclusion rules of the cohort
}
\description{
Get inclusion rules from cohort expression and prepare R language
}
