% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelBuildLangFn.R
\name{getPCCall}
\alias{getPCCall}
\title{Get primary criteria from cohort expression and prepare R language}
\usage{
getPCCall(x)
}
\arguments{
\item{x}{the circe cohort definition}
}
\value{
r language to generate the primary criteria of the cohort
}
\description{
Get primary criteria from cohort expression and prepare R language
}
