% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\docType{class}
\name{Group-class}
\alias{Group-class}
\title{An S4 class for Group}
\description{
TODO clarify the description of a group.
A group that bundles criteria together identifying an event
}
\section{Slots}{

\describe{
\item{\code{Type}}{a expression type class Boolean for the number of items to make the group count}

\item{\code{CriteriaList}}{a list of items (counts and queries) that would identify a medical event}

\item{\code{DemographicCriteriaList}}{a list of demographic attributes that could identify a population}

\item{\code{Groups}}{a list of other groups that are contained within a group}
}}

