% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{gt}
\alias{gt}
\alias{gt,integer-method}
\alias{gt,numeric-method}
\alias{gt,Date-method}
\title{Greater than operator}
\usage{
gt(x)

\S4method{gt}{integer}(x)

\S4method{gt}{numeric}(x)

\S4method{gt}{Date}(x)
}
\arguments{
\item{x}{Comparison bound (integer, numeric, or Date).}
}
\value{
An opAttribute for use inside attribute constructor functions.
}
\description{
Builds a comparison operator for use inside attribute functions such as \code{\link{age}},
\code{\link{valueAsNumber}}, \code{\link{startDate}}, etc.
Accepts integer, numeric, or Date inputs; returns the matching opAttribute type.
}
\examples{
# HbA1c > 6.5
measurement(cs(1L, name = "HbA1c"), valueAsNumber(gt(6.5)))
}
\seealso{
\code{\link{lt}}, \code{\link{gte}}, \code{\link{lte}}, \code{\link{eq}}, \code{\link{bt}}
}
