% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{gte}
\alias{gte}
\alias{gte,integer-method}
\alias{gte,numeric-method}
\alias{gte,Date-method}
\title{Greater than or equal to operator}
\usage{
gte(x)

\S4method{gte}{integer}(x)

\S4method{gte}{numeric}(x)

\S4method{gte}{Date}(x)
}
\arguments{
\item{x}{Comparison bound (integer, numeric, or Date).}
}
\value{
An opAttribute for use inside attribute constructor functions.
}
\description{
Builds a comparison operator for use inside attribute functions such as \code{\link{age}},
\code{\link{valueAsNumber}}, \code{\link{startDate}}, etc.
Accepts integer, numeric, or Date inputs; returns the matching opAttribute type.
}
\examples{
# Age 18 or older
conditionOccurrence(cs(1L, name = "test"), age(gte(18L)))
}
\seealso{
\code{\link{lt}}, \code{\link{gt}}, \code{\link{lte}}, \code{\link{eq}}, \code{\link{bt}}
}
