% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\name{initialize,CensorWindow-method}
\alias{initialize,CensorWindow-method}
\title{Initialization function for s4 cCensorWindow}
\usage{
\S4method{initialize}{CensorWindow}(.Object, StartDate = NA_character_, EndDate = NA_character_)
}
\arguments{
\item{.Object}{an object to initialize}

\item{StartDate}{NA character}

\item{EndDate}{NA character}
}
\value{
an initial CensorWindow object
}
\description{
Initialization function for s4 cCensorWindow
}
