% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelClasses.R
\name{initialize,CollapseSettings-method}
\alias{initialize,CollapseSettings-method}
\title{Initialization function for s4 CollapseSettings}
\usage{
\S4method{initialize}{CollapseSettings}(.Object, CollapseType = "ERA", EraPad = 0L)
}
\arguments{
\item{.Object}{an object to initialize}

\item{CollapseType}{default character string ERA}

\item{EraPad}{default integer 0}
}
\value{
an initial CollapseSettings object
}
\description{
Initialization function for s4 CollapseSettings
}
