% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelClasses.R
\name{initialize,ConceptSetExpression-method}
\alias{initialize,ConceptSetExpression-method}
\title{Initialization function for s4 conceptSetExpression}
\usage{
\S4method{initialize}{ConceptSetExpression}(
  .Object,
  id = uuid::UUIDgenerate(),
  Name = NA_character_,
  Expression = list()
)
}
\arguments{
\item{.Object}{an object to initialize}

\item{id}{generate a guid for the new concept set expression instance}

\item{Name}{empty character name}

\item{Expression}{empty list}
}
\value{
an initial ConceptSetExpression object with a guid
}
\description{
A basic structure to initialize conceptSetExpression. start an id with a guid, no name and an empty list
}
