% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\name{initialize,ConceptSetItem-method}
\alias{initialize,ConceptSetItem-method}
\title{Initialization function for s4 conceptSetItem}
\usage{
\S4method{initialize}{ConceptSetItem}(
  .Object,
  Concept = new("Concept"),
  isExcluded = FALSE,
  includeDescendants = TRUE,
  includeMapped = FALSE
)
}
\arguments{
\item{.Object}{an object to initialize}

\item{Concept}{new concept class}

\item{isExcluded}{default FALSE}

\item{includeDescendants}{default TRUE}

\item{includeMapped}{default FALS}
}
\value{
an initial ConceptSetItem object
}
\description{
A basic structure to initialize conceptSetItem
}
