% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelClasses.R
\name{initialize,EndOfCtsObsEndStrategy-method}
\alias{initialize,EndOfCtsObsEndStrategy-method}
\title{Initialization function for s4 "EndOfCtsObsEndStrategy}
\usage{
\S4method{initialize}{EndOfCtsObsEndStrategy}(.Object, EndOfContinuousObservation = TRUE)
}
\arguments{
\item{.Object}{an object to initialize}

\item{EndOfContinuousObservation}{set TRUE}
}
\value{
an initial end strategy object
}
\description{
Initialization function for s4 "EndOfCtsObsEndStrategy
}
