% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\name{initialize,ExpressionType-method}
\alias{initialize,ExpressionType-method}
\title{Initialization function for s4 classes}
\usage{
\S4method{initialize}{ExpressionType}(.Object, Type = "ALL", Count = NA_integer_)
}
\arguments{
\item{.Object}{an object to initialize}

\item{Type}{default ALL expressions}

\item{Count}{NA_integer}
}
\value{
an initial expressionType object
}
\description{
A basic structure to initialize objects
}
