% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\name{initialize,Query-method}
\alias{initialize,Query-method}
\title{Initialization function for s4 Query}
\usage{
\S4method{initialize}{Query}(
  .Object,
  Domain = NA_character_,
  CodesetId = NA_character_,
  Attributes = list()
)
}
\arguments{
\item{.Object}{an object to initialize}

\item{Domain}{NA character string}

\item{CodesetId}{NA character string}

\item{Attributes}{null list}
}
\value{
an initial ConceptSetItem object
}
\description{
A basic structure to initialize a query with anull domain, codesetid and empty list
}
