% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isStandard.R
\name{isStandard}
\alias{isStandard}
\title{Check for Non-Standard Concepts in Database Tables}
\usage{
isStandard(db_connection, data_concepts_path, vocab_schema, save_path = NULL)
}
\arguments{
\item{db_connection}{A DBI database connection object to the SQL database containing the standard concepts table.}

\item{data_concepts_path}{The file path to the directory containing CSV files of tables to be checked against the standard concepts table. Each CSV file should contain at least the columns `sourceCode` and `concept_id`.}

\item{vocab_schema}{The schema name of the vocabulary database containing the standard concepts table.}

\item{save_path}{(Optional) The file path where the joined tables containing non-standard concepts should be saved. If not provided, the tables will not be saved but will still be checked for non-standard concepts.}
}
\value{
A tibble containing the columns `concept_id`, `concept_name`, `source_code`, and `source_table`, which represent the concept ID, concept name, source code, and source table name for each non-standard concept found.
}
\description{
This function examines a set of tables for non-standard concepts by comparing them against a standard concepts table in a SQL database. It identifies non-standard concepts and optionally saves the joined tables that contain these non-standard concepts.
}
\details{
The function first queries the `cdm.concept` table from the provided SQL database to retrieve the standard concepts. It then reads each CSV file in the specified directory, preparing and joining it with the standard concepts table based on the concept ID. Non-standard concepts are identified based on the `standard_concept` column not being 'S' or 'C'. Information about these non-standard concepts is collected and returned as a tibble.

If a `save_path` is provided, each joined table that contains at least one non-standard concept is saved to the specified directory with the same name as the original table file.
}
\examples{
# Assuming you have a valid DBI connection `db_conn` and your tables are located in "path/to/data_concepts":
non_standard_concepts <- isStandard(db_conn, "path/to/data_concepts", "path/to/save_non_standard/")

}
