% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\docType{class}
\name{Limit-class}
\alias{Limit-class}
\title{An S4 class for Limit}
\description{
A class designating a limit of events per person Types include: all first last
}
\section{Slots}{

\describe{
\item{\code{Type}}{how to limit events per person: all, first, or last}
}}

