% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lineBreak}
\alias{lineBreak}
\title{Print a line break}
\usage{
lineBreak(t = c(1, 2, 3, 4))
}
\arguments{
\item{t}{A number from 1 to 4 representing the type of line break}
}
\value{
Prints a line break. Does not return a value.
}
\description{
Print a line break
}
