% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userCreateAttributeFn.R
\name{listAttributeOptions}
\alias{listAttributeOptions}
\title{List Attribute options}
\usage{
listAttributeOptions(domain = NULL)
}
\arguments{
\item{domain}{the attribute options within the domain, default is NULL then all options printed}
}
\value{
A dataframe with the list of options for attributes we can use specified per domain.
}
\description{
List Attribute options
}
