% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCommands.R
\name{loadComponent}
\alias{loadComponent}
\title{Function to load component}
\usage{
loadComponent(path)
}
\arguments{
\item{path}{a path to the file we wish to load}
}
\value{
returns a component
}
\description{
This function loads the component from a json file to its s4 componentclass
}
