% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelClasses.R
\docType{class}
\name{LogicAttribute-class}
\alias{LogicAttribute-class}
\title{An S4 class for Logic Attribute}
\description{
This class creates a logic attribute which says either true or false if the name of the attribute
is maintained
}
\section{Slots}{

\describe{
\item{\code{Name}}{a name of the attribute}

\item{\code{Logic}}{TRUE or FALSE for this attribute}
}}

