% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserConceptLookupFn.R
\name{lookupConceptCodes}
\alias{lookupConceptCodes}
\title{Lookup Concepts by OMOP Concept Code using Vocabulary}
\usage{
lookupConceptCodes(
  vocabulary,
  conceptCode,
  cdmDatabaseSchema = NULL,
  mapToStandard = TRUE
)
}
\arguments{
\item{vocabulary}{source vocabulary to search (i.e. ICD10, SNOMED). Must be character string}

\item{conceptCode}{source code of vocabulary to serach (example: if ICD10 E11 is T2D). Must be a character string.}

\item{cdmDatabaseSchema}{designate cdm Database schema, if connected to cdm then leave NULL}

\item{mapToStandard}{logic toggle to map the concepts to standard OMOP concepts}
}
\value{
a data frame is returned ordered:
concept_id, concept_name, standard_concept, standard_concept_caption
invalid_reason, invalid_reason_caption, concept_code, domain_id, vocabulary_id, concept_class_id.
}
\description{
This function looks up concepts using the OMOP concept code and vocabulary. Function requires a dbms connection to use
}
