% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userConceptLookupFn.R
\name{lookupConceptIds}
\alias{lookupConceptIds}
\title{Lookup Concepts by OMOP Concept Id}
\usage{
lookupConceptIds(conceptIds, cdmDatabaseSchema = NULL, mapToStandard = TRUE)
}
\arguments{
\item{conceptIds}{standard concept id}

\item{cdmDatabaseSchema}{designate cdm Database schema, if connected to cdm then leave NULL}

\item{mapToStandard}{logic toggle to map the concepts to standard OMOP concepts}
}
\value{
a data frame is returned ordered:
concept_id, concept_name, standard_concept, standard_concept_caption
invalid_reason, invalid_reason_caption, concept_code, domain_id, vocabulary_id, concept_class_id.
}
\description{
This function looks up concepts using the OMOP concept id. Function requires a dbms connection to use
}
