% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{lotNumber}
\alias{lotNumber}
\title{Lot number attribute for DrugExposure}
\usage{
lotNumber(op)
}
\arguments{
\item{op}{an opAttribute object (character) that defines the logical operation and value
(e.g. \code{contains("LOT")} for lot numbers containing "LOT")}
}
\value{
An attribute for use in \code{\link{drugExposure}()}
}
\description{
Filter drug exposure criteria by lot number (e.g. \code{contains("LOT123")} for lot numbers containing "LOT123").
Used only in a drug exposure query.
}
