% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{lt}
\alias{lt}
\alias{lt,integer-method}
\alias{lt,numeric-method}
\alias{lt,Date-method}
\title{Less than operator}
\usage{
lt(x)

\S4method{lt}{integer}(x)

\S4method{lt}{numeric}(x)

\S4method{lt}{Date}(x)
}
\arguments{
\item{x}{Comparison bound (integer, numeric, or Date).}
}
\value{
An opAttribute for use inside attribute constructor functions.
}
\description{
Builds a comparison operator for use inside attribute functions such as \code{\link{age}},
\code{\link{valueAsNumber}}, \code{\link{startDate}}, etc.
Accepts integer, numeric, or Date inputs; returns the matching opAttribute type.
}
\examples{
# Patients under 18
conditionOccurrence(cs(1L, name = "test"), age(lt(18L)))
}
\seealso{
\code{\link{gt}}, \code{\link{gte}}, \code{\link{lte}}, \code{\link{eq}},
  \code{\link{bt}}, \code{\link{age}}, \code{\link{valueAsNumber}}
}
