% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{lte}
\alias{lte}
\alias{lte,integer-method}
\alias{lte,numeric-method}
\alias{lte,Date-method}
\title{Less than or equal to operator}
\usage{
lte(x)

\S4method{lte}{integer}(x)

\S4method{lte}{numeric}(x)

\S4method{lte}{Date}(x)
}
\arguments{
\item{x}{Comparison bound (integer, numeric, or Date).}
}
\value{
An opAttribute for use inside attribute constructor functions.
}
\description{
Builds a comparison operator for use inside attribute functions such as \code{\link{age}},
\code{\link{valueAsNumber}}, \code{\link{startDate}}, etc.
Accepts integer, numeric, or Date inputs; returns the matching opAttribute type.
}
\examples{
# Age 65 or younger
conditionOccurrence(cs(1L, name = "test"), age(lte(65L)))
}
\seealso{
\code{\link{lt}}, \code{\link{gt}}, \code{\link{gte}}, \code{\link{eq}}, \code{\link{bt}}
}
