% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserConceptLookupFn.R
\name{mapConceptToStandard}
\alias{mapConceptToStandard}
\title{Map to a Standard Concept}
\usage{
mapConceptToStandard(
  conceptsDf = NULL,
  conceptId = NULL,
  cdmDatabaseSchema = NULL
)
}
\arguments{
\item{conceptsDf}{a dataframe containing a concept id that can be mapped to standard. Use this in a pipe}

\item{conceptId}{a non-standard concept Id used to map to a standard}

\item{cdmDatabaseSchema}{designate cdm Database schema, if connected to cdm then leave NULL}
}
\value{
a dataframe containing that mapped standard concept id
}
\description{
This function allows you to map a non-standard concept to a standard concept.
Necessary in order to run valid queries in the OMOP CDM
}
