% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelUtilityFn.R
\name{mapOperator}
\alias{mapOperator}
\title{map the operator among options}
\usage{
mapOperator(op)
}
\arguments{
\item{op}{the operator input we want to map}
}
\value{
the circe op
}
\description{
map the operator among options
}
