% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSet.R
\name{mapped}
\alias{mapped}
\title{Include mapped concepts in a concept set}
\usage{
mapped(...)
}
\arguments{
\item{...}{One or more numeric vectors that can be coerced to integers, or
Calls to helper functions "exclude", "descendants", or "mapped".}
}
\value{
A list of Capr concepts
}
\description{
`mapped` is meant to be used inside `cs` when creating a new concept set.
}
