% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{measurement}
\alias{measurement}
\title{Query the measurement domain}
\usage{
measurement(conceptSet, ...)
}
\arguments{
\item{conceptSet}{A measurement concept set (optional)}

\item{...}{optional attributes}
}
\value{
A Capr Query
}
\description{
Query the measurement domain
}
