% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{measurementSourceConcept}
\alias{measurementSourceConcept}
\title{Add a measurement source concept attribute}
\usage{
measurementSourceConcept(conceptSet)
}
\arguments{
\item{conceptSet}{a ConceptSet object containing the source concepts}
}
\value{
An attribute that can be used in a measurement query
}
\description{
Add a measurement source concept attribute
}
