% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\docType{class}
\name{MetaData-class}
\alias{MetaData-class}
\title{An S4 class for Meta Data}
\description{
A class for meta data, info about component structure
}
\section{Slots}{

\describe{
\item{\code{ComponentClass}}{name of component class (this is formally defined)}

\item{\code{Name}}{name for component customized by user}

\item{\code{Description}}{description of the component}

\item{\code{Index}}{A character string either IndexStartDate or IndexEndDate Identifies where the index is relative to the window}
}}

