% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{nbt}
\alias{nbt}
\alias{nbt,integer-method}
\alias{nbt,numeric-method}
\alias{nbt,Date-method}
\title{Not between operator}
\usage{
nbt(x, y)

\S4method{nbt}{integer}(x, y)

\S4method{nbt}{numeric}(x, y)

\S4method{nbt}{Date}(x, y)
}
\arguments{
\item{x}{Lower bound (integer, numeric, or Date).}

\item{y}{Upper bound; must match type of \code{x}.}
}
\value{
An opAttribute for use inside attribute constructor functions.
}
\description{
Builds an exclusion range comparison for use inside attribute functions.
Accepts integer, numeric, or Date inputs.
}
\examples{
# Age outside 18-65
conditionOccurrence(cs(1L, name = "test"), age(nbt(18L, 65L)))
}
\seealso{
\code{\link{bt}}, \code{\link{lt}}, \code{\link{gt}}
}
