% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-nested.R
\docType{class}
\name{nestedAttribute-class}
\alias{nestedAttribute-class}
\title{An S4 class for a nested attribute}
\description{
An S4 class for a nested attribute
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the attribute}

\item{\code{conceptSet}}{a list representing the concepts for the attribute}
}}

