% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-nested.R
\name{nestedWithAll}
\alias{nestedWithAll}
\title{Function to construct a nested group where all criteria and groups must be satisfied}
\usage{
nestedWithAll(...)
}
\arguments{
\item{...}{a set of criteria or groups}
}
\description{
Function to construct a nested group where all criteria and groups must be satisfied
}
