% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-nested.R
\name{nestedWithAtMost}
\alias{nestedWithAtMost}
\title{Function to construct a nested group where at most some of the criteria or groups must be satisfied}
\usage{
nestedWithAtMost(x, ...)
}
\arguments{
\item{x}{an integer specifying the number of criteria or groups that must be satisfied}

\item{...}{a set of criteria or groups}
}
\description{
Function to construct a nested group where at most some of the criteria or groups must be satisfied
}
