% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{notContains}
\alias{notContains}
\alias{notContains,character-method}
\title{String does not contain operator}
\usage{
notContains(x)

\S4method{notContains}{character}(x)
}
\arguments{
\item{x}{the substring to NOT search for}
}
\description{
function that builds an opAttribute based on negated substring containment logic
}
