% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{notEndsWith}
\alias{notEndsWith}
\alias{notEndsWith,character-method}
\title{String does not end with operator}
\usage{
notEndsWith(x)

\S4method{notEndsWith}{character}(x)
}
\arguments{
\item{x}{the suffix string to NOT match}
}
\description{
function that builds an opAttribute based on negated string suffix logic
}
