% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{notStartsWith}
\alias{notStartsWith}
\alias{notStartsWith,character-method}
\title{String does not start with operator}
\usage{
notStartsWith(x)

\S4method{notStartsWith}{character}(x)
}
\arguments{
\item{x}{the prefix string to NOT match}
}
\description{
function that builds an opAttribute based on negated string prefix logic
}
