% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit.R
\docType{class}
\name{ObservationExit-class}
\alias{ObservationExit-class}
\title{An S4 class for a cohort exit based on end of continuous observation.}
\description{
The event persists until the end of continuous observation of the persons
}
\section{Slots}{

\describe{
\item{\code{index}}{specification of event date to offset}

\item{\code{offsetDays}}{an integer specifying the number of days to offset from the event date}
}}

