% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{observationPeriod}
\alias{observationPeriod}
\title{Query the observation period domain}
\usage{
observationPeriod(...)
}
\arguments{
\item{...}{optional attributes}
}
\value{
A Capr Query of domain observation period
}
\description{
Query the observation period domain
}
