% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-conceptSetSelection.R
\name{observationQualifierCS}
\alias{observationQualifierCS}
\title{Observation qualifier filter via concept set}
\usage{
observationQualifierCS(conceptSet, isExclusion = FALSE)
}
\arguments{
\item{conceptSet}{A \code{ConceptSet} built with \code{cs()}.}

\item{isExclusion}{logical; if \code{TRUE} the types in the concept set are excluded.}
}
\value{
An attribute for use in \code{observation()}.
}
\description{
Observation qualifier filter via concept set
}
