% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{observationSourceConcept}
\alias{observationSourceConcept}
\title{Add a observation source concept attribute}
\usage{
observationSourceConcept(conceptSet)
}
\arguments{
\item{conceptSet}{a ConceptSet object containing the source concepts}
}
\value{
An attribute that can be used in a query function
}
\description{
Add a observation source concept attribute
}
