% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-conceptSetSelection.R
\name{observationTypeCS}
\alias{observationTypeCS}
\title{ObservationType filter via concept set (ConceptSetSelection)}
\usage{
observationTypeCS(conceptSet, isExclusion = FALSE)
}
\arguments{
\item{conceptSet}{A \code{ConceptSet} built with \code{cs()}.}

\item{isExclusion}{logical; if \code{TRUE} the types in the concept set are excluded.}
}
\value{
An attribute for use in \code{observation()}.
}
\description{
ObservationType filter via concept set (ConceptSetSelection)
}
