% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-logic.R
\name{observationTypeExclude}
\alias{observationTypeExclude}
\title{ObservationTypeExclude query attribute}
\usage{
observationTypeExclude(exclude = FALSE)
}
\arguments{
\item{exclude}{logical; FALSE = include all types (default), TRUE = exclude}
}
\value{
An attribute for use in \code{observation()}
}
\description{
ObservationTypeExclude query attribute
}
