% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\docType{class}
\name{ObservationWindow-class}
\alias{ObservationWindow-class}
\title{An S4 class for ObservationWindow}
\description{
A class designating an amount of time necessary for an initial event to be recorded
}
\section{Slots}{

\describe{
\item{\code{PriorDays}}{minimal amount of time before event for it to be recorded}

\item{\code{PostDays}}{minimal amount of time after an event for it to be recorded}
}}

