% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{observationWindow}
\alias{observationWindow}
\title{A function to construct the observationWindow}
\usage{
observationWindow(priorDays = 0L, postDays = 0L)
}
\arguments{
\item{priorDays}{minimum number of days prior to the cohort index. Default 0 days}

\item{postDays}{minimum number of days post cohort index. Default 0 days}
}
\description{
A function to construct the observationWindow
}
