% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\docType{class}
\name{Occurrence-class}
\alias{Occurrence-class}
\title{An S4 class for an occurrence.}
\description{
This determines how many events need to occur to count the criteria
in the cohort definition (relative to the index event)
}
\section{Slots}{

\describe{
\item{\code{type}}{a character string determine the logic for counting occurrences. Can be
all, any, exactly, atLeast, or atMost}

\item{\code{count}}{an integer specifying the number of occurrences for a criteria}

\item{\code{isDistinct}}{optional logical; if TRUE, count distinct (e.g. by CountColumn)}

\item{\code{countColumn}}{optional character; when counting, use this column for distinct (e.g. "DOMAIN_CONCEPT")}
}}

