% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{occurrenceCount}
\alias{occurrenceCount}
\title{Occurrence count attribute for ConditionEra}
\usage{
occurrenceCount(op)
}
\arguments{
\item{op}{an opAttribute object (integer) that defines the logical operation and value
(e.g. \code{eq(0L)} for "count equals 0")}
}
\value{
An attribute for use in \code{\link{conditionEra}()}
}
\description{
Filter condition era criteria by the number of condition eras (e.g. occurrence count = 0
for no matching eras). Used only in a condition era query.
}
