% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{offset}
\alias{offset}
\title{A function to offset the number of days relative to index}
\usage{
offset(days)
}
\arguments{
\item{days}{a number specifying the number of days to offset from index where
an event may be observed. In this function a negative number means days before index
and a postive number means days after index.}
}
\description{
A function to offset the number of days relative to index
}
