% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\docType{class}
\name{OpAttribute-class}
\alias{OpAttribute-class}
\title{An S4 class for an Op Attribute}
\description{
An operator attribute meaning it has some value with a boolean operator
}
\section{Slots}{

\describe{
\item{\code{Name}}{the name of the attribute}

\item{\code{Op}}{the operator gt,lt,gte,lte,eq,neq,bt,!bt}

\item{\code{Contents}}{the contents of the attribute as a list. includes the value and the extent}
}}

