% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{class}
\name{opAttributeCharacter-class}
\alias{opAttributeCharacter-class}
\title{An S4 class for a op attribute that is a character}
\description{
An S4 class for a op attribute that is a character
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the attribute}

\item{\code{op}}{the operator one of: startsWith, contains, endsWith, !startsWith, !contains, !endsWith}

\item{\code{value}}{a value or pattern for comparison}

\item{\code{extent}}{unused for character attributes}
}}

