% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{class}
\name{opAttributeNumeric-class}
\alias{opAttributeNumeric-class}
\title{An S4 class for a op attribute that is a numeric}
\description{
An S4 class for a op attribute that is a numeric
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the attribute}

\item{\code{op}}{the operator one of: gt,lt,gte,lte,eq,bt,!bt}

\item{\code{value}}{a value serving as the single limit or lower limit in a bt}

\item{\code{extent}}{a value serving as the upper limit in a bt, otherwise this is empty}
}}

