% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{class}
\name{opAttributeSuper-class}
\alias{opAttributeSuper-class}
\title{An S4 super class for other opAttribute objects to inherit.}
\description{
An S4 super class for other opAttribute objects to inherit.
}
