% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{payerConcept}
\alias{payerConcept}
\title{Payer concept attribute for PayerPlanPeriod}
\usage{
payerConcept(conceptSet)
}
\arguments{
\item{conceptSet}{A ConceptSet for the payer concept filter.}
}
\value{
An attribute for use in \code{payerPlanPeriod()}.
}
\description{
Payer concept attribute for PayerPlanPeriod
}
