% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{payerSourceConcept}
\alias{payerSourceConcept}
\title{Payer source concept attribute for PayerPlanPeriod}
\usage{
payerSourceConcept(conceptSet)
}
\arguments{
\item{conceptSet}{A ConceptSet for the payer source concept filter.}
}
\value{
An attribute for use in \code{payerPlanPeriod()}.
}
\description{
Payer source concept attribute for PayerPlanPeriod
}
