% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{periodLength}
\alias{periodLength}
\title{Period length attribute — ObservationPeriod, PayerPlanPeriod}
\usage{
periodLength(op)
}
\arguments{
\item{op}{an opAttribute object (integer or numeric) defining the length comparison in days.}
}
\value{
An attribute for use in \code{observationPeriod()} or \code{payerPlanPeriod()}.
}
\description{
Period length attribute — ObservationPeriod, PayerPlanPeriod
}
