% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-conceptSetSelection.R
\name{periodTypeCS}
\alias{periodTypeCS}
\title{ObservationPeriod period type filter via concept set (ConceptSetSelection)}
\usage{
periodTypeCS(conceptSet, isExclusion = FALSE)
}
\arguments{
\item{conceptSet}{A \code{ConceptSet} built with \code{cs()}.}

\item{isExclusion}{logical; if \code{TRUE} the types in the concept set are excluded.}
}
\value{
An attribute for use in \code{observationPeriod()}.
}
\description{
ObservationPeriod period type filter via concept set (ConceptSetSelection)
}
