% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-conceptSetSelection.R
\name{placeOfServiceCS}
\alias{placeOfServiceCS}
\title{Place of service filter via concept set — VisitOccurrence, VisitDetail}
\usage{
placeOfServiceCS(conceptSet, isExclusion = FALSE)
}
\arguments{
\item{conceptSet}{A \code{ConceptSet} built with \code{cs()}.}

\item{isExclusion}{logical; if \code{TRUE} the types in the concept set are excluded.}
}
\value{
An attribute for use in \code{visit()} or \code{visitDetail()}.
}
\description{
Place of service filter via concept set — VisitOccurrence, VisitDetail
}
