% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{planConcept}
\alias{planConcept}
\title{Plan concept attribute for PayerPlanPeriod}
\usage{
planConcept(conceptSet)
}
\arguments{
\item{conceptSet}{A ConceptSet for the plan concept filter.}
}
\value{
An attribute for use in \code{payerPlanPeriod()}.
}
\description{
Plan concept attribute for PayerPlanPeriod
}
