% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{planSourceConcept}
\alias{planSourceConcept}
\title{Plan source concept attribute for PayerPlanPeriod}
\usage{
planSourceConcept(conceptSet)
}
\arguments{
\item{conceptSet}{A ConceptSet for the plan source concept filter.}
}
\value{
An attribute for use in \code{payerPlanPeriod()}.
}
\description{
Plan source concept attribute for PayerPlanPeriod
}
