% Generated by roxygen2 (4.0.2): do not edit by hand
\name{post_rc}
\alias{post_rc}
\title{Submit an HTML Form to REDCap}
\usage{
post_rc(uri, certificate, verbose = FALSE, ...)
}
\arguments{
\item{uri}{The uri for your REDCap project.}

\item{certificate}{Certificate file (with path) containing REDCap security
certificate.}

\item{verbose}{Logical (default = FALSE). Verbose output from the posting.}

\item{...}{REDCap API inputs to send in the post.}
}
\value{
If inputs supplied, then a Large Character object; else an error.
}
\description{
Wrapper for \code{RCurl::postForm}. Requires a security certificate to work
and the inputs for the REDCap API (it will fail without them).
}

