% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserPrintFn.R
\name{printCapr,Concept-method}
\alias{printCapr,Concept-method}
\alias{printCapr}
\alias{printCapr,ConceptSetItem-method}
\alias{printCapr,ConceptSetExpression-method}
\alias{printCapr,ObservationWindow-method}
\alias{printCapr,Window-method}
\alias{printCapr,Timeline-method}
\alias{printCapr,OpAttribute-method}
\alias{printCapr,Query-method}
\alias{printCapr,Count-method}
\alias{printCapr,Group-method}
\title{Print concept summary}
\usage{
\S4method{printCapr}{Concept}(x)

\S4method{printCapr}{ConceptSetItem}(x)

\S4method{printCapr}{ConceptSetExpression}(x)

\S4method{printCapr}{ObservationWindow}(x)

\S4method{printCapr}{Window}(x)

\S4method{printCapr}{Timeline}(x)

\S4method{printCapr}{OpAttribute}(x)

\S4method{printCapr}{Query}(x)

\S4method{printCapr}{Count}(x)

\S4method{printCapr}{Group}(x)
}
\arguments{
\item{x}{a criteria class object in s4}
}
\value{
a rpinted summary of the component info
}
\description{
These function print a summary of the CAPR s4 objects
}
