% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserPrintFn.R
\name{printComponent}
\alias{printComponent}
\title{Show Contents of a Component}
\usage{
printComponent(x, showFullConceptSetExpressions = FALSE)
}
\arguments{
\item{x}{the component s4 class you want to preview}

\item{showFullConceptSetExpressions}{T/F options to include full details of concept expressions}
}
\value{
A console print summarizing the s4 component part
}
\description{
This function prints the contents of a component. Note 1/27/21 attributes and some other s4 classes
need to be implemented
}
