% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{procedure}
\alias{procedure}
\title{Query the procedure domain}
\usage{
procedure(conceptSet, ...)
}
\arguments{
\item{conceptSet}{A procedure concept set (optional)}

\item{...}{optional attributes}
}
\value{
A Capr Query
}
\description{
Query the procedure domain
}
