% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{procedureSourceConcept}
\alias{procedureSourceConcept}
\title{Add a procedure source concept attribute}
\usage{
procedureSourceConcept(conceptSet)
}
\arguments{
\item{conceptSet}{a ConceptSet object containing the source concepts}
}
\value{
An attribute that can be used in a query function
}
\description{
Add a procedure source concept attribute
}
