% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-logic.R
\name{procedureTypeExclude}
\alias{procedureTypeExclude}
\title{ProcedureTypeExclude query attribute}
\usage{
procedureTypeExclude(exclude = FALSE)
}
\arguments{
\item{exclude}{logical; FALSE = include all types (default), TRUE = exclude}
}
\value{
An attribute for use in \code{procedure()}
}
\description{
ProcedureTypeExclude query attribute
}
