% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-conceptSetSelection.R
\name{providerSpecialtyCS}
\alias{providerSpecialtyCS}
\title{Provider specialty filter via concept set — multiple domains}
\usage{
providerSpecialtyCS(conceptSet, isExclusion = FALSE)
}
\arguments{
\item{conceptSet}{A \code{ConceptSet} built with \code{cs()}.}

\item{isExclusion}{logical; if \code{TRUE} the types in the concept set are excluded.}
}
\value{
An attribute for use in query functions that support provider specialty.
}
\description{
Provider specialty filter via concept set — multiple domains
}
