% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{quantityValue}
\alias{quantityValue}
\title{Quantity attribute — ProcedureOccurrence, DeviceExposure, Specimen}
\usage{
quantityValue(op)
}
\arguments{
\item{op}{an opAttribute object (numeric or integer) defining the quantity comparison.}
}
\value{
An attribute for use in \code{procedure()}, \code{deviceExposure()}, or \code{specimen()}.
}
\description{
Quantity attribute — ProcedureOccurrence, DeviceExposure, Specimen
}
