% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\docType{class}
\name{Query-class}
\alias{Query-class}
\title{An S4 class for a Circe Query}
\description{
A query is a medical concept that can be extracted from a database through a 'where' clause in a SQL statement.
This includes concepts.
}
\section{Slots}{

\describe{
\item{\code{domain}}{The domain to search (e.g. "Condition", "Drug", "Measurement", etc)}

\item{\code{conceptSet}}{The Concept set describing the observation to serach for}

\item{\code{attributes}}{a list of attributes that modify the query (e.g. `male()`, `female()`, `age(gte(65))`)}
}}

