% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{rangeHighRatio}
\alias{rangeHighRatio}
\title{Range high ratio attribute for Measurement}
\usage{
rangeHighRatio(op)
}
\arguments{
\item{op}{an opAttribute object (numeric) that defines the logical operation and value}
}
\value{
An attribute for use in \code{\link{measurement}()}
}
\description{
Filter measurement criteria by the ratio value_as_number / range_high (e.g. \code{gt(2)} for ratio > 2).
Used only in a measurement query.
}
