% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{rangeLow}
\alias{rangeLow}
\title{Function to create rangeLow attribute}
\usage{
rangeLow(op)
}
\arguments{
\item{op}{an opAttribute object that is either numeric or integer that defines
the logical operation used to determine eligible range low}
}
\description{
This function is used only for measurement query. range_low is a column
in the measurement table of the cdm. This attribute allows a subquery to find measurements
that satisfy certain values determined by the op logic.
}
