% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rc_data}
\alias{rc_data}
\title{Export REDCap Data (Records) from REDCap}
\usage{
rc_data(url, certificate, token, verbose = FALSE, format = "csv",
  type = "eav", raw_or_label = "raw", ids = NULL, fields = NULL,
  forms = NULL, events = NULL, friendly = TRUE, ...)
}
\arguments{
\item{url}{The url for your REDCap project.}

\item{certificate}{Certificate file (with path) containing REDCap security}

\item{token}{API token (character string).}

\item{verbose}{Logical (default = FALSE). Verbose output from the posting.}

\item{format}{Currently only handles "csv"}

\item{type}{Can use "eav" (which is long format) or "flat" (which is wide format).}

\item{raw_or_label}{Use either "raw" or "label" for the kind of data you want}

\item{ids}{Character vector of the record ids you want}

\item{fields}{Character vector of the fields you want}

\item{forms}{Character vector of the forms you want}

\item{events}{Character vector of the events you want}

\item{friendly}{Logical (default=TRUE); Should the results be converted to a
data frame.}

\item{...}{Extra inputs to pass to REDCap (see REDCap API documentation)}
}
\value{
If friendly = TRUE, then a data frame containing the requested data.
  If friendly = FALSE then a Large Character object containing the data
  (you probably don't want this).
}
\description{
Talk to REDCap and get your data (records).
}

