% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rc_parse.R
\name{rc_parse}
\alias{rc_parse}
\title{Parse Data Frame t0 REDCap Format}
\usage{
rc_parse(rc_data)
}
\arguments{
\item{rc_data}{Your REDCap data that has been conditoned appropriately.}
}
\value{
a Large Character object containing the data frame in the new
  "Bizarro" format required for a REDCap import.
}
\description{
REDCap needs a "Bizarro" format for reimporting records. This puts it in that
format. Note that this was taken from
https://github.com/etb/my-R-code/blob/master/R-pull-and-push-from-and-to-REDCap.R
}

