% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSet.R
\name{readConceptSet}
\alias{readConceptSet}
\title{Read a concept set json or csv into R}
\usage{
readConceptSet(path, name)
}
\arguments{
\item{path}{Name of concept set file to read in csv or json format. (e.g. "concepts.json")}

\item{name}{the name of the concept set}
}
\description{
Concept sets can be serialized to json or csv file formats. `readConceptSet`
reads the files into R as Capr concepts sets.
}
\examples{
library(Capr)
path <- tempfile("concepts", fileext = ".json")
concepts <- cs(1, 2, descendants(4, 5), exclude(descendants(6, 7)), name = "test")
writeConceptSet(concepts, path = path)

concepts <- readConceptSet(path)

}
