% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCommands.R
\name{readInCirce}
\alias{readInCirce}
\title{Function to read in a circe json}
\usage{
readInCirce(
  jsonPath,
  connectionDetails,
  vocabularyDatabaseSchema = NULL,
  oracleTempSchema = NULL,
  returnHash = FALSE
)
}
\arguments{
\item{jsonPath}{a path to the file we wish to import}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Cannot be left NULL in this function.}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'vocabulary.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
privileges for storing temporary tables.}

\item{returnHash}{if true returns a has table with all components necessary to build the
cohort definition including the cohort definition}

}
\value{
returns the cohort definition
}
\description{
This function reads a circe json an builds the cohort definition in an execution space
}
