% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelUtilityFn.R
\name{removeDupCSE}
\alias{removeDupCSE}
\title{Function that removes duplicate concept set expressions}
\usage{
removeDupCSE(cse)
}
\arguments{
\item{cse}{the list of concept set expressions used in the object}
}
\value{
a list of concept set expressions free of duplicates
}
\description{
Function that removes duplicate concept set expressions
}
