% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{rules}
\alias{rules}
\title{Create a cohort attrition object}
\usage{
rules(..., expressionLimit = c("First", "All", "Last"))
}
\arguments{
\item{...}{Capr groups}

\item{expressionLimit}{how to limit initial events per person either First, All, or Last}
}
\description{
Create a cohort attrition object
}
