% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCommands.R
\name{saveComponent}
\alias{saveComponent}
\title{Function to save component}
\usage{
saveComponent(x, saveName, savePath = getwd())
}
\arguments{
\item{x}{the component you wish to save}

\item{saveName}{a name for the function you want to save}

\item{savePath}{a path to a file to save. Default is the active working directory}
}
\value{
no return in r. json file written to a save point
}
\description{
This function saves the component as a json file. The component is converted from s4 to s3 to
fit the jsonlite function
}
