% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelSaveFn.R
\name{saveState,Concept-method}
\alias{saveState,Concept-method}
\alias{saveState}
\alias{saveState,ConceptSetItem-method}
\alias{saveState,ConceptSetExpression-method}
\alias{saveState,OpAttribute-method}
\alias{saveState,SourceConceptAttribute-method}
\alias{saveState,ConceptAttribute-method}
\alias{saveState,CorrelatedCriteriaAttribute-method}
\alias{saveState,LogicAttribute-method}
\alias{saveState,Window-method}
\alias{saveState,Timeline-method}
\alias{saveState,Occurrence-method}
\alias{saveState,ExpressionType-method}
\alias{saveState,ObservationWindow-method}
\alias{saveState,Limit-method}
\alias{saveState,Query-method}
\alias{saveState,Count-method}
\alias{saveState,Group-method}
\alias{saveState,MetaData-method}
\alias{saveState,DateOffsetEndStrategy-method}
\alias{saveState,CustomEraEndStrategy-method}
\alias{saveState,EndOfCtsObsEndStrategy-method}
\alias{saveState,CollapseSettings-method}
\alias{saveState,CensorWindow-method}
\alias{saveState,Component-method}
\title{Save State for components}
\usage{
\S4method{saveState}{Concept}(x)

\S4method{saveState}{ConceptSetItem}(x)

\S4method{saveState}{ConceptSetExpression}(x)

\S4method{saveState}{OpAttribute}(x)

\S4method{saveState}{SourceConceptAttribute}(x)

\S4method{saveState}{ConceptAttribute}(x)

\S4method{saveState}{CorrelatedCriteriaAttribute}(x)

\S4method{saveState}{LogicAttribute}(x)

\S4method{saveState}{Window}(x)

\S4method{saveState}{Timeline}(x)

\S4method{saveState}{Occurrence}(x)

\S4method{saveState}{ExpressionType}(x)

\S4method{saveState}{ObservationWindow}(x)

\S4method{saveState}{Limit}(x)

\S4method{saveState}{Query}(x)

\S4method{saveState}{Count}(x)

\S4method{saveState}{Group}(x)

\S4method{saveState}{MetaData}(x)

\S4method{saveState}{DateOffsetEndStrategy}(x)

\S4method{saveState}{CustomEraEndStrategy}(x)

\S4method{saveState}{EndOfCtsObsEndStrategy}(x)

\S4method{saveState}{CollapseSettings}(x)

\S4method{saveState}{CensorWindow}(x)

\S4method{saveState}{Component}(x)
}
\arguments{
\item{x}{a criteria class object in s4}
}
\value{
the object converted to s3 to be saved as a json object
}
\description{
These function coerce s4 CAPR objects to s3 so that they are in a json save state
}
