% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\name{show,Window-method}
\alias{show,Window-method}
\alias{show}
\alias{show,Timeline-method}
\alias{show,Occurrence-method}
\alias{show,ObservationWindow-method}
\alias{show,Group-method}
\alias{show,Query-method}
\alias{show,Count-method}
\alias{show,OpAttribute-method}
\alias{show,Concept-method}
\alias{show,ConceptSetItem-method}
\alias{show,ConceptSetExpression-method}
\alias{show,Limit-method}
\alias{show,MetaData-method}
\alias{show,Component-method}
\alias{show,CohortDetails-method}
\alias{show,CohortDefinition-method}
\title{Show statements of capr objects}
\usage{
\S4method{show}{Window}(object)

\S4method{show}{Timeline}(object)

\S4method{show}{Occurrence}(object)

\S4method{show}{ObservationWindow}(object)

\S4method{show}{Group}(object)

\S4method{show}{Query}(object)

\S4method{show}{Count}(object)

\S4method{show}{OpAttribute}(object)

\S4method{show}{Concept}(object)

\S4method{show}{ConceptSetItem}(object)

\S4method{show}{ConceptSetExpression}(object)

\S4method{show}{Limit}(object)

\S4method{show}{MetaData}(object)

\S4method{show}{Component}(object)

\S4method{show}{CohortDetails}(object)

\S4method{show}{CohortDefinition}(object)
}
\arguments{
\item{object}{the object to show}
}
\value{
a console print of the object
}
\description{
These functions print the capr object to console in a readable format
}
