% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelClasses.R
\docType{class}
\name{SourceConceptAttribute-class}
\alias{SourceConceptAttribute-class}
\title{An S4 class for SourceConceptAttribute}
\description{
An attribute that looks at utilizing the source concepts instead of standard concepts
}
\section{Slots}{

\describe{
\item{\code{Name}}{name of the attribute}

\item{\code{SourceCodesetId}}{a source concept id, conection to concept set expression}
}}

