% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{specimen}
\alias{specimen}
\title{Query the specimen domain}
\usage{
specimen(conceptSet, ...)
}
\arguments{
\item{conceptSet}{A specimen concept set}

\item{...}{optional attributes (e.g. CorrelatedCriteria)}
}
\value{
A Capr Query
}
\description{
Query the specimen domain
}
