% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-conceptSetSelection.R
\name{specimenAnatomicSiteCS}
\alias{specimenAnatomicSiteCS}
\title{Specimen anatomic site filter via concept set}
\usage{
specimenAnatomicSiteCS(conceptSet, isExclusion = FALSE)
}
\arguments{
\item{conceptSet}{A \code{ConceptSet} built with \code{cs()}.}

\item{isExclusion}{logical; if \code{TRUE} the types in the concept set are excluded.}
}
\value{
An attribute for use in \code{specimen()}.
}
\description{
Specimen anatomic site filter via concept set
}
