% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{specimenSourceId}
\alias{specimenSourceId}
\title{Filter by specimen source ID text — Specimen}
\usage{
specimenSourceId(op)
}
\arguments{
\item{op}{an \code{opAttributeCharacter} built with e.g. \code{stringContains("src")}.}
}
\value{
An attribute for use in \code{specimen()}.
}
\description{
Filter by specimen source ID text — Specimen
}
