% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{sponsorConcept}
\alias{sponsorConcept}
\title{Sponsor concept attribute for PayerPlanPeriod}
\usage{
sponsorConcept(conceptSet)
}
\arguments{
\item{conceptSet}{A ConceptSet for the sponsor concept filter.}
}
\value{
An attribute for use in \code{payerPlanPeriod()}.
}
\description{
Sponsor concept attribute for PayerPlanPeriod
}
