% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{sponsorSourceConcept}
\alias{sponsorSourceConcept}
\title{Sponsor source concept attribute for PayerPlanPeriod}
\usage{
sponsorSourceConcept(conceptSet)
}
\arguments{
\item{conceptSet}{A ConceptSet for the sponsor source concept filter.}
}
\value{
An attribute for use in \code{payerPlanPeriod()}.
}
\description{
Sponsor source concept attribute for PayerPlanPeriod
}
