% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{startsWith}
\alias{startsWith}
\alias{startsWith,character-method}
\title{String starts with operator}
\usage{
startsWith(x)

\S4method{startsWith}{character}(x)
}
\arguments{
\item{x}{the prefix string to search for}
}
\description{
function that builds an opAttribute based on string prefix logic
}
