% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{stopReason}
\alias{stopReason}
\title{Filter by stop reason text — ConditionOccurrence or DrugExposure}
\usage{
stopReason(op)
}
\arguments{
\item{op}{an \code{opAttributeCharacter} built with e.g. \code{stringContains("reason")}.}
}
\value{
An attribute for use in \code{conditionOccurrence()} or \code{drugExposure()}.
}
\description{
Filter by stop reason text — ConditionOccurrence or DrugExposure
}
