% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{stopReasonConcept}
\alias{stopReasonConcept}
\title{Stop reason concept attribute for PayerPlanPeriod}
\usage{
stopReasonConcept(conceptSet)
}
\arguments{
\item{conceptSet}{A ConceptSet for the stop reason concept filter.}
}
\value{
An attribute for use in \code{payerPlanPeriod()}.
}
\description{
Stop reason concept attribute for PayerPlanPeriod
}
