% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{stopReasonSourceConcept}
\alias{stopReasonSourceConcept}
\title{Stop reason source concept attribute for PayerPlanPeriod}
\usage{
stopReasonSourceConcept(conceptSet)
}
\arguments{
\item{conceptSet}{A ConceptSet for the stop reason source concept filter.}
}
\value{
An attribute for use in \code{payerPlanPeriod()}.
}
\description{
Stop reason source concept attribute for PayerPlanPeriod
}
