% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{stringContains}
\alias{stringContains}
\alias{stringContains,character-method}
\title{String contains operator}
\usage{
stringContains(x)

\S4method{stringContains}{character}(x)
}
\arguments{
\item{x}{the substring to search for}
}
\description{
function that builds an opAttribute based on substring containment logic
}
