% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{stringEndsWith}
\alias{stringEndsWith}
\alias{stringEndsWith,character-method}
\title{String ends with operator}
\usage{
stringEndsWith(x)

\S4method{stringEndsWith}{character}(x)
}
\arguments{
\item{x}{Character suffix to match.}
}
\value{
An opAttributeCharacter for use inside text attribute functions.
}
\description{
String ends with operator
}
\seealso{
\code{\link{stringContains}}, \code{\link{stringStartsWith}}, \code{\link{lotNumber}}
}
