% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{stringNotContains}
\alias{stringNotContains}
\alias{stringNotContains,character-method}
\title{String does not contain operator}
\usage{
stringNotContains(x)

\S4method{stringNotContains}{character}(x)
}
\arguments{
\item{x}{the substring to NOT search for}
}
\description{
function that builds an opAttribute based on negated substring containment logic
}
