% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{stringNotEndsWith}
\alias{stringNotEndsWith}
\alias{stringNotEndsWith,character-method}
\title{String does not end with operator}
\usage{
stringNotEndsWith(x)

\S4method{stringNotEndsWith}{character}(x)
}
\arguments{
\item{x}{the suffix string to NOT match}
}
\description{
function that builds an opAttribute based on negated string suffix logic
}
