% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{stringNotStartsWith}
\alias{stringNotStartsWith}
\alias{stringNotStartsWith,character-method}
\title{String does not start with operator}
\usage{
stringNotStartsWith(x)

\S4method{stringNotStartsWith}{character}(x)
}
\arguments{
\item{x}{Character prefix to exclude.}
}
\value{
An opAttributeCharacter for use inside text attribute functions.
}
\description{
String does not start with operator
}
\seealso{
\code{\link{stringNotContains}}, \code{\link{stringNotEndsWith}}
}
