% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{stringStartsWith}
\alias{stringStartsWith}
\alias{stringStartsWith,character-method}
\title{String starts with operator}
\usage{
stringStartsWith(x)

\S4method{stringStartsWith}{character}(x)
}
\arguments{
\item{x}{Character prefix to match.}
}
\value{
An opAttributeCharacter for use inside text attribute functions.
}
\description{
String starts with operator
}
\seealso{
\code{\link{stringContains}}, \code{\link{stringEndsWith}}, \code{\link{lotNumber}}
}
