% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelClasses.R
\docType{class}
\name{Timeline-class}
\alias{Timeline-class}
\title{An S4 class for Timeline}
\description{
The timeline class provides context to when the criteria must be observed in a person timeline to pretain to the expression
}
\section{Slots}{

\describe{
\item{\code{StartWindow}}{a window class object identifying the start window}

\item{\code{EndWindow}}{a window class object ifentifying the end window (optional)}

\item{\code{RestrictVisit}}{a logic toggle where TRUE restricts to the same visit}

\item{\code{IgnoreObservationPeriod}}{a logic toggle where TRUE allows events outside the observation period}
}}

