% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{toCirce}
\alias{toCirce}
\title{Function to coerce cohort to circe}
\usage{
toCirce(cd)
}
\arguments{
\item{cd}{the Capr cohort class}
}
\value{
an s3 list representing the circe object to coerce to json
}
\description{
Function to coerce cohort to circe
}
