% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{toConceptSetJson}
\alias{toConceptSetJson}
\title{Convert a Capr concept set to JSON}
\usage{
toConceptSetJson(object, ...)
}
\arguments{
\item{object}{A Capr concept set created with \code{cs()}.}

\item{...}{Arguments passed on to jsonlite::toJSON.}
}
\value{
A JSON character string for the concept set expression.
}
\description{
Convert a Capr concept set to JSON
}
