% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateFn.R
\name{toggleConceptMapping}
\alias{toggleConceptMapping}
\title{Toggle the concept mapping for select positions}
\usage{
toggleConceptMapping(
  conceptMapping,
  pos,
  mapping = c("includeDescendants", "isExcluded", "includeMapped")
)
}
\arguments{
\item{conceptMapping}{the conceptMapping object}

\item{pos}{the positions to toggle}

\item{mapping}{select the mapping type to toggle at each position}
}
\value{
This function returns a list for concept mapping for the concept set expression
}
\description{
This functions changes the logical object (TRUE or FALSE) to its other state. This helps toggle the concept
mapping for a select set in a large list
}
