% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{uniqueDeviceId}
\alias{uniqueDeviceId}
\title{Filter by unique device ID text — DeviceExposure}
\usage{
uniqueDeviceId(op)
}
\arguments{
\item{op}{an \code{opAttributeCharacter} built with e.g. \code{stringContains("DV")}.}
}
\value{
An attribute for use in \code{deviceExposure()}.
}
\description{
Filter by unique device ID text — DeviceExposure
}
