% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{unit}
\alias{unit}
\title{Add unit attribute to a query}
\usage{
unit(...)
}
\arguments{
\item{...}{list of concept ids that identify units}
}
\value{
An attribute that can be used in a query function
}
\description{
Add unit attribute to a query
}
