% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-conceptSetSelection.R
\name{unitCS}
\alias{unitCS}
\title{Unit filter via concept set — Measurement, Observation, DoseEra, Specimen}
\usage{
unitCS(conceptSet, isExclusion = FALSE)
}
\arguments{
\item{conceptSet}{A \code{ConceptSet} built with \code{cs()}.}

\item{isExclusion}{logical; if \code{TRUE} the types in the concept set are excluded.}
}
\value{
An attribute for use in \code{measurement()}, \code{observation()}, \code{doseEra()}, \code{specimen()}.
}
\description{
Unit filter via concept set — Measurement, Observation, DoseEra, Specimen
}
