% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelCoercionFn.R
\name{UpdateAndConvert}
\alias{UpdateAndConvert}
\title{A function to update codeset Ids and convert to circe}
\usage{
UpdateAndConvert(x, conceptTable)
}
\arguments{
\item{x}{the object to update and convert}

\item{conceptTable}{a merge table to match guid to codeset id integer}
}
\value{
an object with updated codeset id
}
\description{
A function to update codeset Ids and convert to circe
}
