% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelCoercionFn.R
\name{UpdateCirceCodesetId,SourceConceptAttribute-method}
\alias{UpdateCirceCodesetId,SourceConceptAttribute-method}
\alias{UpdateCirceCodesetId}
\alias{UpdateCirceCodesetId,Query-method}
\alias{UpdateCirceCodesetId,Count-method}
\alias{UpdateCirceCodesetId,Group-method}
\alias{UpdateCirceCodesetId,CustomEraEndStrategy-method}
\title{Change CodesetId to Integer}
\usage{
\S4method{UpdateCirceCodesetId}{SourceConceptAttribute}(x, conceptTable)

\S4method{UpdateCirceCodesetId}{Query}(x, conceptTable)

\S4method{UpdateCirceCodesetId}{Count}(x, conceptTable)

\S4method{UpdateCirceCodesetId}{Group}(x, conceptTable)

\S4method{UpdateCirceCodesetId}{CustomEraEndStrategy}(x, conceptTable)
}
\arguments{
\item{x}{a component class object in s4}

\item{conceptTable}{a merge table to match guid to codeset id integer}
}
\value{
an object with updated codeset id
}
\description{
When creating the circe json object, an internal reference system needs to be established for the concept set expressions.
This function will update the concept ids from its guid to the ordering of the ids in a merge table. The codeset Ids
will be integers starting from 0 in the circe instance.
}
