% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelCoercionFn.R
\name{UpdateCodesetIdRule}
\alias{UpdateCodesetIdRule}
\title{Update codeset id for inclusion rule}
\usage{
UpdateCodesetIdRule(x, conceptTable)
}
\arguments{
\item{x}{the group that need to update codeset Ids}

\item{conceptTable}{a merge table to match guid to codeset id integer}
}
\value{
an object with updated codeset id
}
\description{
Update codeset id for inclusion rule
}
