% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{valueAsNumber}
\alias{valueAsNumber}
\title{Function to create valueAsNumber attribute}
\usage{
valueAsNumber(op)
}
\arguments{
\item{op}{an opAttribute object that is either numeric or integer that defines the logical
operation used to determine eligible patient age}
}
\value{
An attribute that can be used in a cohort definition
}
\description{
This function is used only for measurement query. valueAsNumber is a column in the measurement
table of the cdm. This attribute allows a subquery to find measurements that satisfy certain values
determined by the op logic.
}
