% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{visitDetail}
\alias{visitDetail}
\title{Query the visit detail domain}
\usage{
visitDetail(conceptSet, ...)
}
\arguments{
\item{conceptSet}{A visit detail concept set}

\item{...}{optional attributes (e.g. VisitDetailSourceConcept)}
}
\value{
A Capr Query
}
\description{
Query the visit detail domain
}
