% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{visitDetailSourceConcept}
\alias{visitDetailSourceConcept}
\title{Add a visit detail source concept attribute}
\usage{
visitDetailSourceConcept(conceptSet)
}
\arguments{
\item{conceptSet}{a ConceptSet object containing the source concepts}
}
\value{
An attribute for use in visitDetail()
}
\description{
Add a visit detail source concept attribute
}
