% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-conceptSetSelection.R
\name{visitDetailTypeCS}
\alias{visitDetailTypeCS}
\title{Visit detail type filter via concept set — VisitDetail}
\usage{
visitDetailTypeCS(conceptSet, isExclusion = FALSE)
}
\arguments{
\item{conceptSet}{A \code{ConceptSet} built with \code{cs()}.}

\item{isExclusion}{logical; if \code{TRUE} the types in the concept set are excluded.}
}
\value{
An attribute for use in \code{visitDetail()}.
}
\description{
Visit detail type filter via concept set — VisitDetail
}
