% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{visitLength}
\alias{visitLength}
\title{Visit length attribute for VisitOccurrence}
\usage{
visitLength(op)
}
\arguments{
\item{op}{an opAttribute object (numeric or integer) defining the length comparison in days}
}
\value{
An attribute for use in \code{visit()}.
}
\description{
Visit length attribute for VisitOccurrence
}
