% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{visitSourceConcept}
\alias{visitSourceConcept}
\title{Add a visit source concept attribute}
\usage{
visitSourceConcept(conceptSet)
}
\arguments{
\item{conceptSet}{a ConceptSet object containing the source concepts}
}
\value{
An attribute that can be used in a query function
}
\description{
Add a visit source concept attribute
}
