% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-logic.R
\name{visitTypeExclude}
\alias{visitTypeExclude}
\title{VisitTypeExclude query attribute}
\usage{
visitTypeExclude(exclude = FALSE)
}
\arguments{
\item{exclude}{logical; FALSE = include all types (default), TRUE = exclude}
}
\value{
An attribute for use in \code{visit()}
}
\description{
VisitTypeExclude query attribute
}
