% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{visitTypeSet}
\alias{visitTypeSet}
\title{Add a visit type attribute (filter by visit_concept_id) from a concept set}
\usage{
visitTypeSet(conceptSet)
}
\arguments{
\item{conceptSet}{a ConceptSet object containing visit type concepts}
}
\value{
An attribute for use in conditionOccurrence(), drugExposure(), etc.
}
\description{
Restricts criteria to events occurring in specified visit types (e.g. inpatient, ER).
Used when decompiling cohort JSON that has VisitType as concept list or CodesetId.
}
