% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelClasses.R
\docType{class}
\name{Window-class}
\alias{Window-class}
\title{An S4 class for a Window}
\description{
A window class provides details on the end points of the timeline
}
\section{Slots}{

\describe{
\item{\code{Event}}{a character string either EventStarts or EventEnds. Identifies the point of reference for the window}

\item{\code{Start}}{a list containing the days and coefficient for the start of the window}

\item{\code{End}}{A list containing the days and coefficient for the end of the window}

\item{\code{Index}}{A character string either IndexStartDate or IndexEndDate Identifies where the index is relative to the window}
}}

