% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{withAny}
\alias{withAny}
\title{Function to construct a group where any criterias and groups may be satisfied}
\usage{
withAny(...)
}
\arguments{
\item{...}{a set of criteria or groups}
}
\description{
Function to construct a group where any criterias and groups may be satisfied
}
