% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{withAtLeast}
\alias{withAtLeast}
\title{Function to construct a group where at least some of the criteria or groups must be satisfied}
\usage{
withAtLeast(x, ...)
}
\arguments{
\item{x}{an integer specifying the number of criterias or groups that must be satisfied}

\item{...}{a set of criteria or groups}
}
\description{
Function to construct a group where at least some of the criteria or groups must be satisfied
}
