% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{withAtMost}
\alias{withAtMost}
\title{Function to construct a group where at most some of the criteria or groups must be satisfied}
\usage{
withAtMost(x, ...)
}
\arguments{
\item{x}{an integer specifying the number of criteria or groups that must be satisfied}

\item{...}{a set of criteria or groups}
}
\value{
an S4 group class specifying a bundle of criteria that at most x be satisfied in context of the cohort definition
}
\description{
Function to construct a group where at most some of the criteria or groups must be satisfied
}
