% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCommands.R
\name{writeCaprCall}
\alias{writeCaprCall}
\title{Function to write capr calls from a circe json}
\usage{
writeCaprCall(jsonPath, txtPath)
}
\arguments{
\item{jsonPath}{a path to the file we wish to import}

\item{txtPath}{a path to the txt file we wish to save}
}
\value{
no return but saves the CAPR calls to build a cohort in a txt file
}
\description{
This function writes the CAPR calls used to build the cohort definition
defined in the circe json . The ouput is a txt file with executable R language
}
