% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{writeCohort}
\alias{writeCohort}
\title{Write Cohort json file}
\usage{
writeCohort(x, path)
}
\arguments{
\item{x}{A Capr cohort}

\item{path}{The name of the file to create}
}
\value{
Invisibly returns the path to the json file that was written
}
\description{
Write Cohort json file
}
\examples{
\dontrun{
cs1 <- cs(descendants(exclude(436665),440383,442306,4175329), name = "concepts")
# optional step to fill in concept set details. Requires database connection.

cs1 <- getConceptSetDetails(cs1, con)

x <- cohort(conditionOccurrence(cs1))
writeCohort(x, "cohortDefinition.json")
}
}
