% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSet.R
\name{writeConceptSet}
\alias{writeConceptSet}
\title{Save a concept set as a json file}
\usage{
writeConceptSet(x, path, format = "auto", ...)
}
\arguments{
\item{x}{A Capr concept set created by `cs()`}

\item{path}{Name of file to write to. (e.g. "concepts.json")}

\item{format}{the file extension to write}

\item{...}{additional arguments}
}
\description{
The resulting concept Set JSON file can be imported into Atlas.
}
\examples{
\dontrun{
anemia <- cs(descendants(439777,4013073,4013074))
writeConceptSet(anemia, 'anemia.json')
writeConceptSet(anemia, 'anemia.csv')
}
}
