% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{age}
\alias{age}
\title{Function to create age attribute}
\usage{
age(op)
}
\arguments{
\item{op}{an opAttribute object that is either numeric or integer that defines
the logical operation used to determine eligible patient age}
}
\description{
Function to create age attribute
}
