% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{atLeast}
\alias{atLeast}
\title{Require at least N occurrences of a query within a window}
\usage{
atLeast(
  x,
  query,
  aperture = duringInterval(eventStarts(-Inf, Inf)),
  distinct = NA,
  countColumn = NA_character_
)
}
\arguments{
\item{x}{Minimum integer count of required occurrences.}

\item{query}{A \code{Query} object (e.g. from \code{conditionOccurrence()}).}

\item{aperture}{An \code{EventAperture} built with \code{\link{duringInterval}}. Defaults to all time.}

\item{distinct}{If \code{TRUE}, count distinct values of \code{countColumn}.}

\item{countColumn}{Column to use for distinct counting (e.g. \code{"DOMAIN_CONCEPT"}).}
}
\value{
A \code{Criteria} object for use in \code{\link{withAll}}, \code{\link{withAny}}, etc.
}
\description{
Require at least N occurrences of a query within a window
}
\examples{
metformin <- cs(descendants(1503297L), name = "Metformin")
# At least 2 metformin fills in the 365 days before index
atLeast(2, drugExposure(metformin), duringInterval(eventStarts(-365, 0)))
}
\seealso{
\code{\link{exactly}}, \code{\link{atMost}}, \code{\link{duringInterval}}
}
