% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{atLeast}
\alias{atLeast}
\title{Function to enumerate an minimal count of occurrences}
\usage{
atLeast(x)
}
\arguments{
\item{x}{the integer counting the number of occurrences}
}
\description{
Function to enumerate an minimal count of occurrences
}
