% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{atMost}
\alias{atMost}
\title{Function to enumerate a maximum count of occurrences}
\usage{
atMost(
  x,
  query,
  aperture = duringInterval(eventStarts(-Inf, Inf)),
  distinct = NA,
  countColumn = NA_character_
)
}
\arguments{
\item{x}{the integer counting the number of occurrences}

\item{query}{a query object that provides context to the clinical event of interest}

\item{aperture}{an eventAperture object that shows the temporal span where the event
is to be observed relative to the index event}

\item{distinct}{optional logical; if TRUE, count distinct (e.g. by countColumn)}

\item{countColumn}{optional character; when counting, use this column for distinct (e.g. "DOMAIN_CONCEPT")}
}
\value{
A criteria that can be used in a cohort definition specifying satisfaction of at most x instances of a query
}
\description{
Function to enumerate a maximum count of occurrences
}
