% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{atMost}
\alias{atMost}
\title{Function to enumerate a maximum count of occurrences}
\usage{
atMost(x)
}
\arguments{
\item{x}{the integer counting the number of occurrences}
}
\description{
Function to enumerate a maximum count of occurrences
}
