% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{male}
\alias{male}
\alias{female}
\title{Add male attribute to a query}
\usage{
male()

female()
}
\value{
An attribute that can be used in a query function

An attribute that can be used in a query function
}
\description{
Add male attribute to a query

Add female attribute to a query
}
\section{Functions}{
\itemize{
\item \code{male()}: male demographic attribute

\item \code{female()}: female demographic attribute

}}
\examples{
\dontrun{
# Create a cohort of males with Type 1 diabetes
t1dm <- cs(descendants(201254, 435216, 40484648))
t1dm_males <- cohort(condition(t1dm, male()))
}
\dontrun{
# Create a cohort of males with Type 1 diabetes
t1dm <- cs(descendants(201254, 435216, 40484648))
t1dm_females <- cohort(condition(t1dm, female()))
}
}
