% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{attrition}
\alias{attrition}
\title{Create a cohort attrition object}
\usage{
attrition(..., expressionLimit = c("First", "All", "Last"))
}
\arguments{
\item{...}{Capr groups}

\item{expressionLimit}{how to limit initial events per person either First, All, or Last}
}
\value{
A cohort attrition object that can be used in a cohort definition
}
\description{
Create a cohort attrition object
}
