% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{bt}
\alias{bt}
\alias{bt,integer-method}
\alias{bt,numeric-method}
\alias{bt,Date-method}
\title{Between operator}
\usage{
bt(x, y)

\S4method{bt}{integer}(x, y)

\S4method{bt}{numeric}(x, y)

\S4method{bt}{Date}(x, y)
}
\arguments{
\item{x}{the left side bound of the between logic This can either be an integer, numeric, or Date
data type. Different data types will return the appropriate opAttribute type}

\item{y}{the right side bound of the between logic. This can either be an integer, numeric, or
Date data type. Different data types will return the appropriate opAttribute type}
}
\description{
function that builds an opAttribute based on between logic
}
