% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capr-package.R
\docType{package}
\name{capr-package}
\alias{capr-package}
\title{capr package documentation}
\description{
Infrastructure for Covariate Assisted Principal Regression.
}
\author{
\strong{Maintainer}: Xi Luo \email{xi.rossi.luo@gmail.com}

Authors:
\itemize{
  \item Yi Zhao \email{zhaoyi1026@gmail.com}
  \item Brian Caffo \email{bcaffo@gmail.com}
}

}
\keyword{internal}
