% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{cohort}
\alias{cohort}
\title{Function that creates a cohort object}
\usage{
cohort(entry, attrition = NULL, exit = NULL, era = NULL)
}
\arguments{
\item{entry}{the index event of the cohort}

\item{attrition}{rules that restrict the cohort further, developing attrition}

\item{exit}{the event where the person exits the cohort}

\item{era}{Cohort era (collapse) logic created with the `cohortEra` function}
}
\description{
Function that creates a cohort object
}
