% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\docType{class}
\name{CohortDetails-class}
\alias{CohortDetails-class}
\title{An S4 class providing details for the Cohort}
\description{
An S4 class providing details for the Cohort
}
\section{Slots}{

\describe{
\item{\code{Name}}{a name for the cohort}

\item{\code{Description}}{a text field providing an information on the cohort and what it is intended}

\item{\code{Author}}{who created the cohort}

\item{\code{cdmVersionRange}}{the range of cdm versions}
}}

