% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{compile,Cohort-method}
\alias{compile,Cohort-method}
\alias{compile,ConceptSet-method}
\title{Compile a Capr cohort to json}
\usage{
\S4method{compile}{Cohort}(object, ..., includeConceptSets = NULL)

\S4method{compile}{ConceptSet}(object, ...)
}
\arguments{
\item{object}{A Capr Concept Set created with `cs`}

\item{...}{Arguments passed on to jsonlite::toJSON.
e.g. `pretty = TRUE` for nicely formatted json.}
}
\value{
The json representation of Capr cohorts

The json representation of Capr cohorts
}
\description{
Compile a Capr cohort to json

Compile a Capr Concept Set to json
}
\examples{
ch <- cohort(conditionOccurrence(cs(1,2, name = "concepts")))
compile(ch)
}
