% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelClasses.R
\docType{class}
\name{Component-class}
\alias{Component-class}
\title{An S4 class for Component}
\description{
This class is an flexible container to store information about the cohort definition, allowing us to maintain information
in smaller parts that remain relevant in isolation. The structure of circe cohort definition relies on a concept set
table that stores information for queries. In each cohort component an internal reference id is used to maintain
consistency between the expression of the cohort criteria and the actionable concepts. The component container
bundles the concept set expression and the criteria expression into one object that is saveable and inheritable.
Smaller classes are stored within the container and when they are converted into a superior class the component container
is modified but the previous information is kept in tact. A component consists of 4 parts: meta data which stores
the name, description and the componentClass. The componentClass identifies what kind of component one is using. Next
the criteriaExpression stores any information about the deployment of the medical concept. This includes queries, counts,
groups, attributes and other structures that detail the information of the specific component class. The limit
is a section that specifies the limit of entry for person events. Is it the first event, all events or last event for
the criteriaExpression we are interested in observing. Finally the concept set expression holds the concepts relevant
to the criteria expression. The component can be saved as a json file or loaded back into its s4 class.
}
\section{Slots}{

\describe{
\item{\code{MetaData}}{meta information about the object}

\item{\code{CriteriaExpression}}{a list of criteria that is in the object}

\item{\code{Limit}}{a list containing any limits}

\item{\code{ConceptSetExpression}}{a list containing any concept sets}
}}

