% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSet.R
\docType{class}
\name{Concept-class}
\alias{Concept-class}
\title{An S4 class for a single OMOP Concept}
\description{
A concept class contains all the information about the concept from the OMOP voabulary.
}
\section{Slots}{

\describe{
\item{\code{concept_id}}{the OMOP/OHDSI concept ID}

\item{\code{concept_name}}{the name of the concept}

\item{\code{standard_concept}}{whether the concept is standard 'S', classification 'C', or  non-standard NA}

\item{\code{standard_concept_caption}}{Whether the concept is standard full phrase}

\item{\code{invalid_reason}}{Whether the concept is invalid single letter}

\item{\code{invalid_reason_caption}}{whether the concept is invalid standard phrase}

\item{\code{concept_code}}{The original code of the concept from its vocabulary}

\item{\code{domain_id}}{The domain of the concept (e.g. Drug, Condition, Procedure, etc)}

\item{\code{vocabulary_id}}{the name of the vocabulary}

\item{\code{concept_class_id}}{type of concept class}
}}

