% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\docType{class}
\name{conceptAttribute-class}
\alias{conceptAttribute-class}
\title{An S4 class for a concept attribute}
\description{
An S4 class for a concept attribute
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the attribute}

\item{\code{conceptSet}}{a list representing the concepts for the attribute}
}}

