% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{conditionEra}
\alias{conditionEra}
\title{Query the condition era domain}
\usage{
conditionEra(conceptSet, ...)
}
\arguments{
\item{conceptSet}{A \code{ConceptSet} for the condition.}

\item{...}{Optional attributes: \code{firstOccurrence()}, \code{startDate()}, \code{endDate()},
\code{eraLength()}, \code{occurrenceCount()}, \code{ageAtStart()}, \code{ageAtEnd()}, etc.}
}
\value{
A \code{Query} object.
}
\description{
Query the condition era domain
}
\seealso{
\code{\link{conditionOccurrence}}, \code{\link{entry}}
}
