% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{conditionOccurrence}
\alias{conditionOccurrence}
\title{Query the condition domain}
\usage{
conditionOccurrence(conceptSet, ...)
}
\arguments{
\item{conceptSet}{A condition concept set (optional)}

\item{...}{optional attributes}
}
\value{
A Capr Query
}
\description{
Query the condition domain
}
