% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelCoercionFn.R
\name{convertAdditionalCriteriaToCIRCE}
\alias{convertAdditionalCriteriaToCIRCE}
\title{Convert Additional Criteria Component to CIRCE}
\usage{
convertAdditionalCriteriaToCIRCE(x)
}
\arguments{
\item{x}{the component to convert}
}
\value{
a circe converted component
}
\description{
Convert Additional Criteria Component to CIRCE
}
