% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreateAttributeFn.R
\name{createFirstAttribute}
\alias{createFirstAttribute}
\title{create First Attribute}
\usage{
createFirstAttribute(logic = TRUE)
}
\arguments{
\item{logic}{toggle TRUE for first occurence}
}
\value{
a component of attribute class
}
\description{
This function creates a attribute for first occurrence
}
