% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userCreateAttributeFn.R
\name{createProcedureSourceConceptAttribute}
\alias{createProcedureSourceConceptAttribute}
\title{create procedure source concept}
\usage{
createProcedureSourceConceptAttribute(ConceptSetExpression)
}
\arguments{
\item{ConceptSetExpression}{the concepte set expression we wish to deploy as a source concept attribute
This concept set expression should contain source codes, which may be non-standard.}
}
\value{
a source concept attribute component
}
\description{
create procedure source concept
}
