% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSet.R
\name{cs}
\alias{cs}
\title{Create a concept set}
\usage{
cs(..., name = "", id = NULL)
}
\arguments{
\item{...}{One or more numeric vectors that can be coerced to integers, or
Calls to helper functions "exclude", "descendants", or "mapped". Negative
integers will be marked as excluded from the concept set.}

\item{name}{A name for the concept set}

\item{id}{An id for the concept set}
}
\value{
A Capr Concept Set Object
}
\description{
cs is used to create concept set expressions.
}
\examples{
\dontrun{
cs(1, 2)
cs(1, c(1, 10, 2))
cs(1, seq(2, 10, 2))
cs(1, 2, 3, exclude(4, 5))
cs(1, 2, 3, exclude(4, 5), mapped(6, 7))
cs(1, 2, 3, exclude(4, 5), mapped(6, 7), descendants(8, 9))
cs(descendants(1, 2, 3),  exclude(descendants(8, 9)))
}
}
