% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-dateAdjustment.R
\name{dateAdjustment}
\alias{dateAdjustment}
\title{Function to create age attribute}
\usage{
dateAdjustment(
  startWith = "START_DATE",
  startOffset = 0L,
  endWith = "END_DATE",
  endOffset = 0L
)
}
\arguments{
\item{startWith}{character string either START_DATE or END_DATE}

\item{startOffset}{an integer value, default 0}

\item{endWith}{character string either START_DATE or END_DATE}

\item{endOffset}{an integer value, default 0}
}
\value{
A dateAdjustment attribute class that can be used with a query
}
\description{
Function to create age attribute
}
