% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{death}
\alias{death}
\title{Query the death domain}
\usage{
death(conceptSet = NULL, ...)
}
\arguments{
\item{conceptSet}{A death domain, with rare exceptions this is always NULL}

\item{...}{optional attributes}
}
\value{
A Capr Query
}
\description{
Query the death domain
}
