% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{death}
\alias{death}
\title{Query the condition era domain}
\usage{
death(conceptSet = NULL, ...)
}
\arguments{
\item{conceptSet}{A condition concept set}

\item{...}{optional attributes}
}
\value{
A Capr Query
}
\description{
Query the condition era domain
}
