% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{drugEra}
\alias{drugEra}
\title{Query the drug era domain}
\usage{
drugEra(conceptSet, ...)
}
\arguments{
\item{conceptSet}{A \code{ConceptSet} for the drug ingredient.}

\item{...}{Optional attributes: \code{firstOccurrence()}, \code{startDate()}, \code{endDate()},
\code{eraLength()}, \code{occurrenceCount()}, \code{ageAtStart()}, \code{ageAtEnd()}, etc.}
}
\value{
A \code{Query} object.
}
\description{
Query the drug era domain
}
\seealso{
\code{\link{drugExposure}}, \code{\link{drugExit}}, \code{\link{entry}}
}
