% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{drugEra}
\alias{drugEra}
\title{Query the drug era domain}
\usage{
drugEra(conceptSet, ...)
}
\arguments{
\item{conceptSet}{A drug ingredient concept set}

\item{...}{optional attributes}
}
\value{
A Capr Query
}
\description{
Query the drug era domain
}
