% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{endDate}
\alias{endDate}
\title{Function that creates a end date attribute}
\usage{
endDate(op)
}
\arguments{
\item{op}{an opAttribute object must be a date that defines the logical operation used to
determine eligible end dates}
}
\description{
Function that creates a end date attribute
}
