% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{entry}
\alias{entry}
\title{Create a cohort entry criteria}
\usage{
entry(
  ...,
  observationWindow = observeWindow(),
  primaryCriteriaLimit = "First",
  additionalCriteria = NULL,
  qualifiedLimit = "First"
)
}
\arguments{
\item{...}{Capr Queries}

\item{observationWindow}{a time specifying the minimal time a person is observed}

\item{primaryCriteriaLimit}{Which primary criteria matches should be considered for inclusion? "First", "Last" or "All"}

\item{additionalCriteria}{a Capr group that adds restriction to the entry event}

\item{qualifiedLimit}{Which criteria matches should be considered for inclusion? "First", "Last" or "All"}
}
\value{
A cohort entry Capr object
}
\description{
Create a cohort entry criteria
}
