% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{eq}
\alias{eq}
\alias{eq,integer-method}
\alias{eq,numeric-method}
\alias{eq,Date-method}
\alias{eq,character-method}
\title{Equal to operator}
\usage{
eq(x)

\S4method{eq}{integer}(x)

\S4method{eq}{numeric}(x)

\S4method{eq}{Date}(x)

\S4method{eq}{character}(x)
}
\arguments{
\item{x}{Comparison value (integer, numeric, Date, or character).}
}
\value{
An opAttribute for use inside attribute constructor functions.
}
\description{
Builds a comparison operator for use inside attribute functions such as \code{\link{age}},
\code{\link{valueAsNumber}}, \code{\link{startDate}}, etc.
Accepts integer, numeric, Date, or character inputs.
}
\examples{
# Exactly 2 refills
drugExposure(cs(1L, name = "drug"), drugRefills(eq(2L)))
}
\seealso{
\code{\link{lt}}, \code{\link{gt}}, \code{\link{lte}}, \code{\link{gte}}, \code{\link{bt}}
}
