% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{exactly}
\alias{exactly}
\title{Function to enumerate an exact count of occurrences}
\usage{
exactly(x)
}
\arguments{
\item{x}{the integer counting the number of occurrences}
}
\description{
Function to enumerate an exact count of occurrences
}
