% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{exit}
\alias{exit}
\title{Function that creates a cohort exit object}
\usage{
exit(endStrategy, censor = NULL)
}
\arguments{
\item{endStrategy}{the endStrategy object to specify for the exit}

\item{censor}{the censoring criteria to specify for the exit}
}
\description{
Function that creates a cohort exit object
}
