% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit.R
\name{fixedExit}
\alias{fixedExit}
\title{Function to create an exit based on exit based on the end of a continuous drug exposure}
\usage{
fixedExit(index = c("startDate", "endDate"), offsetDays)
}
\arguments{
\item{index}{specification of event date to offset. Can be either startDate or endDate}

\item{offsetDays}{an integer specifying the number of days to offset from the event date}
}
\description{
Function to create an exit based on exit based on the end of a continuous drug exposure
}
