% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelBuildLangFn.R
\name{getCohortEraCall}
\alias{getCohortEraCall}
\title{Get cohort era from cohort expression and prepare R language}
\usage{
getCohortEraCall(x)
}
\arguments{
\item{x}{the circe cohort definition}
}
\value{
r language to generate the cohort era of the cohort
}
\description{
Get cohort era from cohort expression and prepare R language
}
