% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSet.R
\name{getConceptSetCall}
\alias{getConceptSetCall}
\title{Create the Capr code to build a concept set}
\usage{
getConceptSetCall(x, name = x@Name)
}
\arguments{
\item{x}{A concept set}

\item{name}{the name of the concept set}
}
\value{
The Capr code required to build the concept set
}
\description{
Create the Capr code to build a concept set
}
