% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSet.R
\name{getConceptSetDetails}
\alias{getConceptSetDetails}
\title{Fill in Concept Set details using a vocab}
\usage{
getConceptSetDetails(x, con, vocabularyDatabaseSchema = NULL)
}
\arguments{
\item{x}{A concept set created by `cs`}

\item{con}{A connection to an OMOP CDM database}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary format resides. Note that
for SQL Server, this should include both the database and schema
name, for example 'vocabulary.dbo'.}
}
\value{
A modified version of the input concept set with concept details filled in.
}
\description{
Concept sets created in R using the `cs` function do not contain details like
"CONCEPT_NAME", "DOMAIN_ID", etc. If an OMOP CDM vocabulary is available then
these details can be filled in by the the `getConceptSetDetails` function.
}
\examples{
\dontrun{
# create a concept set
vocabularyDatabaseSchema = "cdm5"
anemia <- cs(descendants(439777,4013073,4013074))

# fill in the details from an OMOP CDM
library(DatabaseConnector)
con <- connect(dbms = "postgresql", user = "postgres", password = "", server = "localhost/covid")
anemia <- getConceptSetDetails(condition_anemia, con, vocabularyDatabaseSchema = "cdm5")
}
}
