% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowLevelUtilityFn.R
\name{getConceptSetId,ConceptSetExpression-method}
\alias{getConceptSetId,ConceptSetExpression-method}
\alias{getConceptSetId}
\title{Function to find the ConceptSetId}
\usage{
\S4method{getConceptSetId}{ConceptSetExpression}(x)
}
\arguments{
\item{x}{the component to check}
}
\value{
the id from the conceptset expression
}
\description{
Function to find the ConceptSetId
}
