% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\docType{class}
\name{Group-class}
\alias{Group-class}
\title{An S4 class for a group}
\description{
a group is the combination of multiple criteria or sub groups
}
\section{Slots}{

\describe{
\item{\code{occurrence}}{an occurrence object specifying how many events must occur
to consider the event as part of the cohort definition}

\item{\code{critera}}{a list of criteria that are grouped together}

\item{\code{group}}{a list of sub-groups to consider}
}}

