% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{gt}
\alias{gt}
\alias{gt,integer-method}
\alias{gt,numeric-method}
\alias{gt,Date-method}
\title{Greater than operator}
\usage{
gt(x)

\S4method{gt}{integer}(x)

\S4method{gt}{numeric}(x)

\S4method{gt}{Date}(x)
}
\arguments{
\item{x}{the value to used as a bound in the op logic. This can either be an integer,
numeric, or Date data type. Different data types will return the appropriate opAttribute type}
}
\description{
function that builds an opAttribute based on greater than logic
}
