% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{gte}
\alias{gte}
\alias{gte,integer-method}
\alias{gte,numeric-method}
\alias{gte,Date-method}
\title{Greater than or equal to operator}
\usage{
gte(x)

\S4method{gte}{integer}(x)

\S4method{gte}{numeric}(x)

\S4method{gte}{Date}(x)
}
\arguments{
\item{x}{the value to used as a bound in the op logic. This can either be an integer,
numeric, or Date data type. Different data types will return the appropriate opAttribute type}
}
\description{
function that builds an opAttribute based on greater than or equal to logic
}
