% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-logic.R
\docType{class}
\name{logicAttribute-class}
\alias{logicAttribute-class}
\title{An S4 class for a logical attribute}
\description{
with a logic attribute if it is specified than we assume it is true
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the attribute}
}}

