% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userConceptLookupFn.R
\name{lookupKeyword}
\alias{lookupKeyword}
\title{Lookup concept name as a general search}
\usage{
lookupKeyword(
  keyword,
  search_type = c("like", "exact", "any"),
  cdmDatabaseSchema = NULL
)
}
\arguments{
\item{keyword}{a word a or phrase to search concepts}

\item{search_type}{how to use keyword: a) like the keyword, b)exact keyword , or c) any match of keyword}

\item{cdmDatabaseSchema}{designate cdm Database schema, if connected to cdm then leave NULL}
}
\value{
a data.table with all concepts found from the search
}
\description{
This function looks up concepts based on the concept name. It can be modified to conduct
an exact name search or general search that contains the concept name in the concept.
}
