% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{lt}
\alias{lt}
\alias{lt,integer-method}
\alias{lt,numeric-method}
\alias{lt,Date-method}
\title{Less than operator}
\usage{
lt(x)

\S4method{lt}{integer}(x)

\S4method{lt}{numeric}(x)

\S4method{lt}{Date}(x)
}
\arguments{
\item{x}{the value to used as a bound in the op logic. This can either be an integer, numeric, or
Date data type. Different data types will return the appropriate opAttribute type}
}
\description{
function that builds an opAttribute based on less than logic
}
