% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{lte}
\alias{lte}
\alias{lte,integer-method}
\alias{lte,numeric-method}
\alias{lte,Date-method}
\title{Less than or equal to operator}
\usage{
lte(x)

\S4method{lte}{integer}(x)

\S4method{lte}{numeric}(x)

\S4method{lte}{Date}(x)
}
\arguments{
\item{x}{the value to used as a bound in the op logic. This can either be an integer,
numeric, or Date data type. Different data types will return the appropriate opAttribute type}
}
\description{
function that builds an opAttribute based on less than or equal to than logic
}
