% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{makeCohortSet}
\alias{makeCohortSet}
\title{Make a cohort dataframe for cohort generator}
\usage{
makeCohortSet(...)
}
\arguments{
\item{...}{multiple capr cohorts to bind into a dataframe}
}
\description{
Make a cohort dataframe for cohort generator
}
