% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{measurementUnit}
\alias{measurementUnit}
\title{Add unit attribute to a query}
\usage{
measurementUnit(x)
}
\arguments{
\item{x}{A concept set that identifies units, built with \code{cs()}. This is the only
supported input type - see \code{getConceptSetDetails()} if you want the concept
set to carry real concept names/domain/vocabulary for display in Atlas.}
}
\value{
An attribute that can be used in a query function
}
\description{
Add unit attribute to a query
}
