% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{measurementUnit}
\alias{measurementUnit}
\title{Add unit attribute to a query}
\usage{
measurementUnit(x)
}
\arguments{
\item{x}{A a concept set that identifies units}
}
\value{
An attribute that can be used in a query function
}
\description{
Add unit attribute to a query
}
