% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{nbt}
\alias{nbt}
\alias{nbt,integer-method}
\alias{nbt,numeric-method}
\alias{nbt,Date-method}
\title{Not between operator}
\usage{
nbt(x, y)

\S4method{nbt}{integer}(x, y)

\S4method{nbt}{numeric}(x, y)

\S4method{nbt}{Date}(x, y)
}
\arguments{
\item{x}{the left side bound of the between logic This can either be an integer,
numeric, or Date data type. Different data types will return the appropriate opAttribute type}

\item{y}{the right side bound of the between logic. This can either be an integer,
numeric, or Date data type. Different data types will return the appropriate opAttribute type}
}
\description{
function that builds an opAttribute based on not between logic
}
