% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-nested.R
\name{nestedWithAny}
\alias{nestedWithAny}
\title{Function to construct a nested group where any criteria and groups may be satisfied}
\usage{
nestedWithAny(...)
}
\arguments{
\item{...}{a set of criteria or groups}
}
\description{
Function to construct a nested group where any criteria and groups may be satisfied
}
