% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{observation}
\alias{observation}
\title{Query the observation domain}
\usage{
observation(conceptSet, ...)
}
\arguments{
\item{conceptSet}{A condition concept set}

\item{...}{optional attributes}
}
\value{
A Capr Query
}
\description{
Query the observation domain
}
