% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit.R
\name{observationExit}
\alias{observationExit}
\title{Function to create an exit based on continuous observation}
\usage{
observationExit()
}
\description{
Function to create an exit based on continuous observation
}
