% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\docType{class}
\name{ObservationWindow-class}
\alias{ObservationWindow-class}
\title{An S4 class for an ObservationWindow}
\description{
this determines the minimal observation time before and after index for all persons
in the cohort
}
\section{Slots}{

\describe{
\item{\code{priorDays}}{minimum number of days prior to the cohort index}

\item{\code{postDays}}{minimum number of days post cohort index}
}}

