% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowLevelClasses.R
\docType{class}
\name{Occurrence-class}
\alias{Occurrence-class}
\title{An S4 class for Occurrence}
\description{
The Occurrence class provides logic on the number of criterias that most be true in a person for them to be contained in
the expression
}
\section{Slots}{

\describe{
\item{\code{Type}}{a character string of either at most, at least, or exactly providing context to the number of occurrences}

\item{\code{Count}}{an integer value that provides the number of occurrences}

\item{\code{isDistinct}}{a logic toggle where if TRUE only counts distinct occurrences}
}}

