% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{payerPlanPeriod}
\alias{payerPlanPeriod}
\title{Query the payer plan period domain (CDM ≥ 5.3)}
\usage{
payerPlanPeriod(...)
}
\arguments{
\item{...}{optional attributes (e.g. \code{firstOccurrence()}, \code{startDate()},
\code{periodLength()}, \code{ageAtStart()}, \code{ageAtEnd()}, \code{genderCS()})}
}
\value{
A Capr Query of domain PayerPlanPeriod
}
\description{
Query the payer plan period domain (CDM ≥ 5.3)
}
