% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rc_codes.R
\name{rc_codes}
\alias{rc_codes}
\title{Create a Code list for REDCap radio, checkbox and yesno fields}
\usage{
rc_codes(dict)
}
\arguments{
\item{dict}{REDCap dictionary (data frame returned by rc_dict).}
}
\value{
A list containing the raw codes and labels for your REDCap radio,
checkbox and yesno fields
}
\description{
Give it your dictionary and it will create a code list
}

