% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rc_dict.R
\name{rc_dict}
\alias{rc_dict}
\title{Export a REDCap Data Dictionary from REDCap}
\usage{
rc_dict(url, token, friendly = TRUE, verbose = FALSE)
}
\arguments{
\item{url}{The url for your REDCap project.}

\item{token}{API token (character string).}

\item{friendly}{Logical (default=TRUE); Should the results be converted to a
data frame.}

\item{verbose}{Logical (default = FALSE). Verbose output from the posting.}
}
\value{
If friendly = TRUE, then a data frame containing the data dictionary.
  If friendly = FALSE then a Large Character object containing the data
  dictionary (you probably don't want this).
}
\description{
Talk to REDCap and get your data dictionary.
}

