% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rc_event}
\alias{rc_event}
\title{Export a REDCap Event Structure from REDCap}
\usage{
rc_event(url, certificate, token, friendly = TRUE, verbose = FALSE)
}
\arguments{
\item{url}{The url for your REDCap project.}

\item{certificate}{Certificate file (with path) containing REDCap security certificate}

\item{token}{API token (character string).}

\item{friendly}{Logical (default=TRUE); Should the results be converted to a
data frame.}

\item{verbose}{Logical (default = FALSE). Verbose output from the posting.}
}
\value{
If friendly = TRUE, then a data frame containing the events.
  If friendly = FALSE then a Large Character object containing the events
  (you probably don't want this).
}
\description{
Talk to REDCap and get your event definitions for longitudinal projects.
}

