% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSet.R
\name{readConceptSet}
\alias{readConceptSet}
\title{Write a concept set to a json file}
\usage{
readConceptSet(path, name, id = NULL)
}
\arguments{
\item{path}{Name of concept set file to read. (e.g. "concepts.json")}
}
\description{
The resulting concept Set JSON file can be imported into Atlas.
}
\examples{
\dontrun{
anemia <- readConceptSet('anemia.json')
}
}
