% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSet.R
\name{readConceptSet}
\alias{readConceptSet}
\title{Read a concept set json or csv into R}
\usage{
readConceptSet(path, name, id = NULL)
}
\arguments{
\item{path}{Name of concept set file to read in csv or json format. (e.g. "concepts.json")}
}
\description{
Concept sets can be serialized to json or csv file formats. `readConceptSet`
reads the files into R as Capr concepts sets.
}
\examples{
\dontrun{
anemia <- readConceptSet('anemia.json')
anemia <- readConceptSet('anemia.csv')
}
}
