% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCommands.R
\name{readInCirce}
\alias{readInCirce}
\title{Function to read in a circe json}
\usage{
readInCirce(jsonPath)
}
\arguments{
\item{jsonPath}{a path to the file we wish to import}
}
\value{
returns the cohort definition
}
\description{
This function reads a circe json an builds the cohort definition in an execution space
}
