% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\name{startDate}
\alias{startDate}
\title{Function that creates a start date attribute}
\usage{
startDate(op)
}
\arguments{
\item{op}{an opAttribute object must be a date that defines
the logical operation used to determine eligible start dates}
}
\description{
Function that creates a start date attribute
}
