% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{stringContains}
\alias{stringContains}
\alias{stringContains,character-method}
\title{String contains operator}
\usage{
stringContains(x)

\S4method{stringContains}{character}(x)
}
\arguments{
\item{x}{Character substring to match.}
}
\value{
An opAttributeCharacter for use inside text attribute functions.
}
\description{
Builds a text-match operator for use with text filter attribute functions such as
\code{\link{lotNumber}}, \code{\link{stopReason}}, \code{\link{uniqueDeviceId}}.
}
\examples{
# Drug exposures where lot number contains "LOT"
drugExposure(cs(1L, name = "drug"), lotNumber(stringContains("LOT")))
}
\seealso{
\code{\link{stringStartsWith}}, \code{\link{stringEndsWith}},
  \code{\link{stringNotContains}}, \code{\link{lotNumber}}, \code{\link{stopReason}}
}
