% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{stringEndsWith}
\alias{stringEndsWith}
\alias{stringEndsWith,character-method}
\title{String ends with operator}
\usage{
stringEndsWith(x)

\S4method{stringEndsWith}{character}(x)
}
\arguments{
\item{x}{the suffix string to search for}
}
\description{
function that builds an opAttribute based on string suffix logic
}
