% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{stringNotContains}
\alias{stringNotContains}
\alias{stringNotContains,character-method}
\title{String does not contain operator}
\usage{
stringNotContains(x)

\S4method{stringNotContains}{character}(x)
}
\arguments{
\item{x}{Character substring to exclude.}
}
\value{
An opAttributeCharacter for use inside text attribute functions.
}
\description{
String does not contain operator
}
\seealso{
\code{\link{stringNotStartsWith}}, \code{\link{stringNotEndsWith}}
}
