% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{stringNotEndsWith}
\alias{stringNotEndsWith}
\alias{stringNotEndsWith,character-method}
\title{String does not end with operator}
\usage{
stringNotEndsWith(x)

\S4method{stringNotEndsWith}{character}(x)
}
\arguments{
\item{x}{Character suffix to exclude.}
}
\value{
An opAttributeCharacter for use inside text attribute functions.
}
\description{
String does not end with operator
}
\seealso{
\code{\link{stringNotStartsWith}}, \code{\link{stringNotContains}}
}
