% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-op.R
\docType{methods}
\name{stringStartsWith}
\alias{stringStartsWith}
\alias{stringStartsWith,character-method}
\title{String starts with operator}
\usage{
stringStartsWith(x)

\S4method{stringStartsWith}{character}(x)
}
\arguments{
\item{x}{the prefix string to search for}
}
\description{
function that builds an opAttribute based on string prefix logic
}
