% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-concept.R
\name{unit}
\alias{unit}
\title{Add unit attribute to a query}
\usage{
unit(x)
}
\arguments{
\item{x}{A single character idetifier for a unit or a concept set that identifies units}
}
\value{
An attribute that can be used in a query function
}
\description{
Add unit attribute to a query
}
\examples{
# create a unit attribute
unit(8713L)
unit("\%")
}
