% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{visit}
\alias{visit}
\title{Query the visit occurrence domain}
\usage{
visit(conceptSet, ...)
}
\arguments{
\item{conceptSet}{A condition concept set (optional)}

\item{...}{optional attributes}
}
\value{
A Capr Query
}
\description{
Query the visit occurrence domain
}
