% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{withAll}
\alias{withAll}
\title{Function to construct a group where all criterias and groups must be satisfied}
\usage{
withAll(...)
}
\arguments{
\item{...}{a set of criteria or groups}
}
\description{
Function to construct a group where all criterias and groups must be satisfied
}
