% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{withAny}
\alias{withAny}
\title{Function to construct a group where any criteria and groups may be satisfied}
\usage{
withAny(...)
}
\arguments{
\item{...}{a set of criteria or groups}
}
\value{
an S4 group class specifying a bundle of criteria that any may be satisfied in context of the cohort definition
}
\description{
Function to construct a group where any criteria and groups may be satisfied
}
