% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{withAny}
\alias{withAny}
\title{Function to construct a group where any criteria and groups may be satisfied}
\usage{
withAny(...)
}
\arguments{
\item{...}{a set of criteria or groups}
}
\description{
Function to construct a group where any criteria and groups may be satisfied
}
