% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{writeCohort}
\alias{writeCohort}
\title{Write Cohort json file}
\usage{
writeCohort(x, path, ...)
}
\arguments{
\item{x}{A Capr cohort}

\item{path}{The name of the file to create}

\item{...}{Further arguments passed on to jsonlite::toJSON}
}
\description{
Write Cohort json file
}
\examples{
\dontrun{
cs1 <- cs(descendants(exclude(436665),440383,442306,4175329))
cs1 <- getConceptSetDetails(cs1)
x <- cohort(condition(cs1))
writeCohort(x, "cohortDefinition.json")
}
}
